### 2018-01-26 #189 @tklebel
Submission: jstor
The tool [Data for Research (DfR)](http://www.jstor.org/dfr/) by JSTOR is a
valuable source for citation analysis and text mining. `jstor`
provides functions and suggests workflows for importing
datasets from DfR.

```
Package: jstor
Title: Read Data from JSTOR/DfR
Version: 0.2.6
Authors@R: person("Thomas", "Klebel", email = "thomas.klebel@uni-graz.at",
                  role = c("aut", "cre"))
Description: Functions and helpers to import metadata and full-texts delivered
    by Data for Research (DfR) by JSTOR.
License: GPL-3 | file LICENSE
    xml2,
    foreach,
    doSNOW,
    snow
Suggests: testthat,
    covr,
    tidyr
```

-   URL for the package (the development repository, not a stylized html page): https://github.com/tklebel/jstor

The package should fit very well into the **data extraction**-category, because
it extracts data from `*.xml`-files for further use.

The target audience would be scientists leveraging JSTOR/DfR for textual research.
Currently there is is no implementation in R to parse the meta data they deliver.
Although the data quality on references is not very consistent, one could try
to conduct citation analysis with a lot of control over the sample and the data
compared to GoogleScholar or Web of Science.

No.

Pre-submission was approved by @karthik: https://github.com/ropensci/onboarding/issues/186

    - [x] The package is deposited in a long-term repository with the DOI: [10.5281/zenodo.1169862](https://zenodo.org/record/1169862#.Wn2VQWaX9QM)
	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

----

I would very much like to submit a paper to JOSS, however at the moment I don't
have a proper case-study yet. If possible, I would like to submit a
paper at a later stage.

----

The only "problems" from `goodpractice::gp()` are a test coverage of 98% and
lengthy lines in the test-files.

If he would find the time to do it, a technical review by @jimhester could
probably help in resolving some performance issues, although he might urge me
to re-write the whole thing (again).
