+++ NEWS.md
@@ -1,3 +1,5 @@
-# jstor 0.2.2.9000
+# jstor 0.2.5.9000
 
+* find_article does not coerce days and months to integer any more, since there
+might be information stored as text.
 * Added a `NEWS.md` file to track changes to the package.
