+++ vignettes/Introduction.Rmd
@@ -93,8 +93,8 @@ The columns are:
 - volume *(chr)*: The volume the article was published in.
 - issue *(chr)*: The issue the article was published in.
 - language *(chr)*: The language of the article.
-- pub_day *(int)*: Publication day, if specified.
-- pub_month *(int)*: Publication month, if specified.
+- pub_day *(chr)*: Publication day, if specified.
+- pub_month *(chr)*: Publication month, if specified.
 - pub_year *(int)*: Year of publication.
 - first_pag *(int)*: Page number for the first page of the article.
 - last_page *(int)*: Page number for the last page of the article.
