### 2018-02-02 377b477
Test for invalid URI, which should not raise a warning #33
