### 2018-02-06 872b0ee
remove argument to extract_basename
