+++ R/full_text.R
@@ -21,7 +21,7 @@ get_encoding <- function(filename) {
 read_full_text <- function(filename) {
   validate_file_path(filename, "txt")
 
-  id <- extract_basename(filename, type = "txt")
+  id <- extract_basename(filename)
 
   encoding <- get_encoding(filename)
 
