### 2018-02-06 b50a1f7
export new helper get_basename
