+++ tests/testthat/test-full_text.R
@@ -5,8 +5,8 @@ crazy_path <- "../Entpackt//soc_a-j/jpalestud/2012/41/2/10.1525%2Fjps.2012.xli.i
 standard_path <- "../Entpackt//soc_k-z/sozialewelt/1952/3/4/i40039236/40874627/40874627.txt"
 
 test_that("getting base name of file works", {
-  expect_identical(extract_basename(crazy_path), "10.1525%2Fjps.2012.xli.2.204")
-  expect_identical(extract_basename(standard_path), "40874627")
+  expect_identical(get_basename(crazy_path), "10.1525%2Fjps.2012.xli.2.204")
+  expect_identical(get_basename(standard_path), "40874627")
 })
 
 # encoding -----
