+++ man/get_basename.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helpers.R
+\name{get_basename}
+\alias{get_basename}
+\title{Extract the basename of a path}
+\usage{
+get_basename(file_path)
+}
+\arguments{
+\item{file_path}{A path to a file}
+}
+\description{
+This helper simply extracts the basename of a path and removes the extension,
+e.g. \code{foo/bar.txt} is shortend to \code{bar}.
+}
