+++ NAMESPACE
@@ -7,6 +7,7 @@ export(find_book)
 export(find_chapters)
 export(find_footnotes)
 export(find_references)
+export(get_basename)
 export(jstor_example)
 export(jstor_import)
 export(read_full_text)
