### 2018-02-06 9d54e2d
update NEWS.md
