+++ NEWS.md
@@ -1,5 +1,7 @@
 # jstor 0.2.5.9000
 
+* new helper `get_basename` helps to get the basename of a file without its
+extension
 * find_article does not coerce days and months to integer any more, since there
 might be information stored as text.
 * Added a `NEWS.md` file to track changes to the package.
