### 2018-02-07 edebb72
correct WS for lintr
