+++ tests/testthat/test-pages.R
@@ -32,4 +32,3 @@ test_that("total pages are computed correctly", {
   expect_identical(find_total_pages(first_page = 2, last_page = 2), 1)
   expect_identical(find_total_pages(first_page = NA, last_page = NA), NA_real_)
 })
-

[…]
