### 2018-02-07 569af6b
default to writing column names
closes #29
