+++ NEWS.md
@@ -1,5 +1,6 @@
 # jstor 0.2.5.9000
 
+* `jstor_import` now writes column names by default #29
 * new helper `get_basename` helps to get the basename of a file without its
 extension
 * find_article does not coerce days and months to integer any more, since there
