### 2018-02-07 3bad96e
change to snow for parallel execution and add progress bars
closes #34
