+++ NAMESPACE
@@ -18,6 +18,7 @@ importFrom(dplyr,mutate)
 importFrom(dplyr,mutate_all)
 importFrom(dplyr,n)
 importFrom(dplyr,rename)
+importFrom(foreach,"%dopar%")
 importFrom(magrittr,"%>%")
 importFrom(purrr,discard)
 importFrom(purrr,flatten_chr)
