+++ man/jstor_import.Rd
@@ -37,3 +37,36 @@ Writes \code{.csv}-files to disk.
 This function applies an import function to a list of \code{xml}-files and saves
 them in batches of \code{.csv}-files to disk.
 }
+\details{
+Along the way, we wrap three functions, which make the process of converting
+many files easier:
+\itemize{
+\item \code{\link[purrr:safely]{purrr::safely()}}
+\item \code{\link[foreach:foreach]{foreach::foreach()}}
+\item \code{\link[readr:write_csv]{readr::write_csv()}}
+}
+
+When using one of the \code{find_*} functions, there should usually be no errors.
+To avoid the whole computation to fail in the unlikely event that an error
+occurs, we use \code{safely()} which let's us
+continue the process, and catch the error along the way.
+
+If you have many files to import, you might benefit from executing the
+function in parallel. We use \code{\link[snow:createCluster]{snow::createCluster()}} to setup a cluster and
+then compute the results via \code{foreach} and \code{\%dopar\%}. The type of
+cluster is determined by \code{getClusterOption("type")}.
+
+After importing all files, they are written to disk with
+\code{\link[readr:write_csv]{readr::write_csv()}}.
+
+Since you might run out of memory when importing a large quantity of files,
+the files to import are split up into batches. Each batch is being treated
+separately, therefore for each batch multiple processes from
+\code{\link[snow:createCluster]{snow::createCluster()}} are
+spawned. For this reason, it is not recommended to have very small batches,
+as there is an overhead for starting and ending the processes. On the other
+hand, the batches should not be too large, to not exceed memory limitations.
+A value of 10000 to 20000 for \code{files_per_batch} should work fine on most
+machines. If the session is interactive and \code{show_progress} is \code{TRUE}, a
+progress bar is displayed for each batch.
+}
