+++ R/batch_import_fun.R
@@ -107,7 +107,9 @@ jstor_convert_to_file <- function(in_paths, chunk_number, out_path, fun,
 #' continue the process, and catch the error along the way.
 #' 
 #' If you have many files to import, you might benefit from executing the
-#' function in parallel. We use [snow::createCluster()] to setup a cluster and
+#' function in parallel. We use
+#' \code{\link[snow:snow-startstop]{snow::createCluster()}} to setup a cluster
+#' and
 #' then compute the results via `foreach` and \code{\%dopar\%}. The type of
 #' cluster is determined by `getClusterOption("type")`.
 #' 
@@ -117,7 +119,7 @@ jstor_convert_to_file <- function(in_paths, chunk_number, out_path, fun,
 #' Since you might run out of memory when importing a large quantity of files,
 #' the files to import are split up into batches. Each batch is being treated
 #' separately, therefore for each batch multiple processes from
-#' [snow::createCluster()] are
+#' \code{\link[snow:snow-startstop]{snow::createCluster()}} are
 #' spawned. For this reason, it is not recommended to have very small batches,
 #' as there is an overhead for starting and ending the processes. On the other
 #' hand, the batches should not be too large, to not exceed memory limitations.
