+++ .travis.yml
@@ -8,7 +8,6 @@ r:
 matrix: 
   include:
   - r: devel
-    warnings_are_errors: false
 
 sudo: false
 cache: packages
