### 2018-02-07 84e3f27
format elapsed time correctly
