### 2018-02-08 c490469
add first draft for JOSS-paper
