+++ paper.bib
@@ -1,0 +1,44 @@
+
+@online{_jstor_2018,
+	title = {{JSTOR}},
+	url = {http://www.jstor.org/},
+	urldate = {2018-02-08},
+	date = {2018},
+	file = {JSTOR:/Users/tklebel/Documents/zotero/storage/VEFHFAU6/www.jstor.org.html:text/html}
+}
+
+@article{roberts_introduction_2016,
+	title = {Introduction to the Virtual Issue: Recent Innovations in Text Analysis for Social Science},
+	volume = {24},
+	issn = {1047-1987, 1476-4989},
+	url = {https://www.cambridge.org/core/journals/political-analysis/article/introduction-to-the-virtual-issue-recent-innovations-in-text-analysis-for-social-science/8CFCA019C25D82254EE0CAC28F3F650E},
+	doi = {10.1017/S1047198700014418},
+	shorttitle = {Introduction to the Virtual Issue},
+	abstract = {Introduction to the Virtual Issue: Recent Innovations in Text Analysis for Social Science - Volume 24 Issue V10 - Margaret E. Roberts},
+	pages = {1--5},
+	issue = {V10},
+	journaltitle = {Political Analysis},
+	author = {Roberts, Margaret E.},
+	urldate = {2018-02-08},
+	date = {2016-04},
+	langid = {english},
+	file = {Full Text PDF:/Users/tklebel/Documents/zotero/storage/BRIWHQBA/Roberts - 2016 - Introduction to the Virtual Issue Recent Innovati.pdf:application/pdf;Snapshot:/Users/tklebel/Documents/zotero/storage/F4PLIM9P/8CFCA019C25D82254EE0CAC28F3F650E.html:text/html}
+}
+
+ @Manual{r_core,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2017},
+    url = {https://www.R-project.org/},
+  }
+  
+  
+  @Manual{jstor,
+    title = {jstor: Read Data from JSTOR/DfR},
+    author = {Thomas Klebel},
+    year = {2018},
+    note = {R package version 0.2.5.9000},
+    url = {https://github.com/tklebel/jstor},
+  }
