+++ paper.md
@@ -4,7 +4,7 @@ authors:
 - affiliation: 1
   name: Thomas Klebel
   orcid: 0000-0002-7331-4751
-date: "14 February 2016"
+date: "2018-02-08"
 output:
   html_document:
     df_print: paged
