### 2018-02-08 f3cf0e7
remove comment about the novelty of jstor from paper
