### 2018-02-08 3420f76
some yaml formatting for paper
