### 2018-02-08 88b70cd
update introductory vignette
