+++ vignettes/Introduction.Rmd
@@ -43,7 +43,7 @@ I will demonstrate their usage using the
 [sample dataset](http://www.jstor.org/dfr/about/sample-datasets)
 which is provided by JSTOR on their website.
 
-## General Concept
+# General Concept
 All functions which are concerned with meta data (therefore excluding
 `find_fulltext`) operate along the same lines:
 
@@ -54,7 +54,7 @@ All functions which are concerned with meta data (therefore excluding
 The following sections showcase the different functions.
 
 
-## Application
+# Application
 Apart from `jstor` we only need to load `dplyr` for matching records.
 
 ```{r, message=FALSE, warning=FALSE}
@@ -64,7 +64,7 @@ library(dplyr)
 
 
 
-### `find_article`
+## `find_article`
 The basic usage of the `find_*` functions is very simple. They take only one 
 argument, the path to the file to import:
 ```{r}
@@ -105,7 +105,7 @@ Since the output from all functions are tibbles, the result is nicely formatted:
 meta_data %>% knitr::kable()
 ```
 
-### `find_authors`
+## `find_authors`
 Extracting the authors works in similar fashion:
 
 ```{r, results='asis'}
@@ -126,7 +126,7 @@ supplied, i.e.: `Albert Einstein`, or `Einstein, Albert`.
 
 The number of rows matches the number of authors -- each author get its' own row.
 
-### `find_references`
+## `find_references`
 ```{r}
 references <- find_references(jstor_example("sample_with_references.xml"))
 ```
@@ -153,7 +153,7 @@ within the article. In the above example there are several headings, because the
 sample file doesn't follow a typical convention (it was published in 1922). 
 
 
-### `find_footnotes`
+## `find_footnotes`
 ```{r}
 find_footnotes(jstor_example("sample_with_references.xml"))
 ```
@@ -171,7 +171,7 @@ footnotes
 
 
 
-### `find_fulltext`
+## `find_fulltext`
 The function to extract full texts can't be demonstrated with proper data, since
 the full texts are only supplied upen special request with DfR. The function
 guesses the encoding of the specified file via `readr::guess_encoding()`, reades
@@ -188,10 +188,10 @@ full_text[["full_text"]]
 ```
 
 
-### Combining results
+## Combining results
 Different parts of meta-data can be combined by using `dplyr::left_join()`.
 
-**Matching with authors**
+###Matching with authors
 
 ```{r, results='asis'}
 meta_data %>% 
@@ -200,18 +200,17 @@ meta_data %>%
   knitr::kable()
 ```
 
-**Matching with references**
+###Matching with references
 
-```{r, results='asis'}
+```{r}
 meta_data %>% 
   left_join(references) %>%
-  select(basename_id, article_title, pub_year) %>%
-  head(10) %>% 
-  knitr::kable()
+  select(basename_id, pub_year, references) %>%
+  head(10) 
 ```
 
 
-## Books
+# Books
 Quite recently DfR added book chapters to their stack. To import metadata about
 the books and chapters, jstor supplies `find_book` and `find_chapters`.
 
