### 2018-02-08 0ab2c5b
spell check docs and vignettes
