### 2018-02-08 bf3e079
update vignette about batch import to reflect recent changes in parallel backend and function params
closes #35
