+++ vignettes/automating_file_import.Rmd
@@ -37,7 +37,7 @@ debugging.
 
 # List files
 There are many ways to generate a list of all files: 
-`list.files` or using `system` in conjunction with `find` on unix-like systems
+`list.files()` or using `system()` in conjunction with `find` on unix-like systems
 are common options. 
 
 For demonstration purposes I use files contained in `jstor` which can be accessed
@@ -99,11 +99,8 @@ size of files. If you try to import many files, you could either run into
 problems regarding RAM, or it would take a long time and would benefit from
 executing the process in parallel.
 
-The function `jstor_convert_to_file` takes care of errors and uses
-`parallel::mclapply` to speed up the process. Since `mclapply` is not supported
-on windows, there is currently no support for parallel processing on windows with
-`jstor_convert_to_file`. Obviously, you could write your own implementation using
-a different engine that works on windows.
+The function `jstor_import()` takes care of errors and uses
+`foreach::foreach()` and `snow::makeCluster()` to speed up the process.
 
 The whole approach about importing many files is implemented like this:
 
@@ -114,8 +111,7 @@ The whole approach about importing many files is implemented like this:
     + write the result into a single .csv-file
     + write any errors in a separate .csv-file
     
-The easiest way to achieve this pipeline is by using `jstor_import_wrapper`. It
-takes up to six arguments:
+`jstor_import()` takes up to seven arguments:
 
 - *in_paths*: this should be a character vector of all `xml-`files you want to
 import
@@ -127,12 +123,14 @@ or `find_footnotes`)
 - *files_per_batch*: how many files should be converted per batch? For my system,
 20,000 files per batch worked well.
 - *cores* number of cores to use for parallel processing (`1` on windows).
+- *show_progress*: if `TRUE` and run from an interactive session, a progress bar
+will be displayed for each batch.
 
 You can specify all parameters like this:
 ```{r, eval=FALSE}
-jstor_import_wrapper(in_paths = vector_of_all_my_files, out_file = "metadata",
-                     out_path = "path/to/out/dir", .f = find_article,
-                     files_per_batch = 20000, cores = 4)
+jstor_import(in_paths = vector_of_all_my_files, out_file = "metadata",
+             out_path = "path/to/out/dir", .f = find_article,
+             files_per_batch = 20000, cores = 4)
 ```
 
 Suppose you had 100,000 files to convert, and no errors happened, then you should
