### 2018-02-08 5f69350
add link to pkgdown to DESCRIPTION
closes #15
