+++ DESCRIPTION
@@ -27,7 +27,7 @@ Suggests: testthat,
     rmarkdown,
     tidyr
 BugReports: https://github.com/tklebel/jstor/issues
-URL: https://github.com/tklebel/jstor
+URL: https://github.com/tklebel/jstor, https://tklebel.github.io/jstor/
 RoxygenNote: 6.0.1
 Roxygen: list(markdown = TRUE)
 VignetteBuilder: knitr
