### 2018-02-08 6d13512
add case-study to pkgdown
