+++ vignettes/articles/analysing-n-grams.Rmd
@@ -1,0 +1,483 @@
+---
+title: Analysing n-grams with jstor for R
+author: "Thomas Klebel"
+date: "`r Sys.Date()`"
+output:
+  rmarkdown::html_vignette:
+      toc: true
+---
+
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+The service [DfR](http://www.jstor.org/dfr/) by JSTOR offers several ways for
+text analysis of scientific articles. 
+In this post I will demonstrate how to analyse n-grams which DfR delivers.
+
+Let's suppose, we are interested in the topic of "inequality" within the 
+discipline of sociology. Social inequality can be considered a prime subject of
+sociological inquiry. In order to gain some context on the subject, we might be
+interested to analyse frequently occurring terms.
+
+Our analysis starts at the main page of DfR. We create a dataset by searching
+for
+"inequality" and selecting "sociology" as our subject. To trim down the number
+of articles, we only select articles from 1997 to 2017. After
+logging in/creating an account, we select unigrams and bigrams. After unzipping
+the archives to a convenient location, we start our analysis.
+
+
+Up-front, we need to load some packages. `jstor` is currently not available from
+CRAN, but can be installed via `devtools`.
+
+
+```{r, message=FALSE}
+# install.packages("devtools")
+# devtools::install_github("tklebel/jstor")
+
+library(jstor)
+library(tidyverse)
+library(visdat)
+
+# set a lighter theme for plots
+theme_set(theme_bw())
+```
+
+To import the files, we first need to locate them and generate an object with
+their corresponding paths. The following code assumes that you follow a workflow
+organised around projects within RStudio (refer to
+http://r4ds.had.co.nz/workflow-projects.html for further information). 
+
+```{r}
+# list files
+meta_files <- list.files(pattern = "xml", full.names = T, recursive = T)
+```
+
+Since we have a decent amount of articles, let's speed up the process of
+importing the metadata via parallel processing. `jstor_import` is a nice wrapper
+which takes care of setting up the processes and deals with any errors that
+might occur along the way.
+
+```{r, eval=FALSE}
+jstor_import(meta_files, out_file = "imported_metadata", .f = find_article,
+             files_per_batch = 25000, cores = 4)
+#> Starting to import 23909 file(s).
+#> Processing chunk 1/1
+#>   |===================================================================| 100%
+#> Finished importing 23909 file(s) in 2.54 mins.
+```
+
+Since `jstor_import` writes the results to disk, we need to read the metadata
+from the newly created file.^[When reading this data without `guess_max = 2000`,
+a warning is raised because the column type was not recognized properly for one
+column. Increasing `guess_max` 
+helped in this instance.]
+
+
+```{r, message=FALSE}
+imported_metadata <- read_csv("imported_metadata-1.csv", guess_max = 2000)
+imported_metadata
+```
+
+
+
+# Exploration
+Before diving into the analysis of ngrams, we might wish to take an explorative
+look at our metadata.
+
+The first thing to look at are the types of articles.
+```{r}
+ggplot(imported_metadata, aes(article_type)) +
+  geom_bar() +
+  coord_flip()
+```
+
+We can see, that the majority of articles are proper "research-articles", which
+together with book-reviews and miscellaneous articles amount to ~99% of all
+articles.
+
+```{r}
+imported_metadata %>% 
+  count(article_type, sort = T) %>% 
+  mutate(perc = scales::percent(n/sum(n)))
+```
+
+We must be cautious, however, when using this variable to distinguish articles
+into categories. In this instance, we have "research-articles" which are
+actually book-reviews:
+
+```{r}
+imported_metadata %>% 
[…]
