+++ _pkgdown.yml
@@ -12,6 +12,8 @@ navbar:
       href: articles/Introduction.html
     - text: Automating File Import
       href: articles/automating_file_import.html
+    - text: Analysing n-grams with jstor
+      href: articles/articles/analysing-n-grams.html
   - text: News
     href: news/index.html
   right:
