### 2018-02-08 40e36da
expand interpretation slightly in case-study
