### 2018-02-08 34f60fa
add badge for rOpenSci review
