### 2018-02-08 c6a6bca
add linkt to personal homepage in pkgdown
