### 2018-02-09 08dff88
add simple helper to suppress warnings when reading xml-files
closes #33
