### 2018-02-09 89e0db3
any is more concise than sum(lgl) > 0
