### 2018-02-09 c98d953
add tables specifying, which fields are supported
closes #20
