### 2018-02-09 52dd955
more NEWS formatting
