+++ NEWS.md
@@ -9,6 +9,6 @@
 * `jstor_import` now writes column names by default #29
 * new helper `get_basename` helps to get the basename of a file without its
 extension
-* find_article does not coerce days and months to integer any more, since there
+* `find_article` does not coerce days and months to integer any more, since there
 might be information stored as text.
 * Added a `NEWS.md` file to track changes to the package.
