+++ README.Rmd
@@ -19,8 +19,8 @@ knitr::opts_chunk$set(
 [![Coverage status](https://codecov.io/gh/tklebel/jstor/branch/master/graph/badge.svg)](https://codecov.io/github/tklebel/jstor?branch=master)
 [![CRAN status](http://www.r-pkg.org/badges/version/jstor)](https://cran.r-project.org/package=jstor)
 [![CRAN\_Download\_Badge](http://cranlogs.r-pkg.org/badges/jstor)](https://CRAN.R-project.org/package=jstor)
-[![](https://badges.ropensci.org/189_status.svg)](https://github.com/ropensci/onboarding/issues/189)
-
+[![rOpenSci badge](https://badges.ropensci.org/189_status.svg)](https://github.com/ropensci/onboarding/issues/189)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1169862.svg)](https://doi.org/10.5281/zenodo.1169862)
   
 The tool [Data for Research (DfR)](http://www.jstor.org/dfr/) by JSTOR is a
 valuable source for citation analysis and text mining. `jstor`
