### 2018-02-11 6b865d8
somem minor issues and commenting
