+++ vignettes/analysing-n-grams.Rmd
@@ -1,5 +1,5 @@
 ---
-title: Analysing n-grams with jstor for R
+title: Analysing n-grams with `jstor` for R
 author: "Thomas Klebel"
 date: "`r Sys.Date()`"
 output:
@@ -19,15 +19,21 @@ In this vignette I will demonstrate how to analyse n-grams which DfR delivers.
 Let's suppose, we are interested in the topic of "inequality" within the 
 discipline of sociology. Social inequality can be considered a prime subject of
 sociological inquiry. In order to gain some context on the subject, we might be
-interested to analyse frequently occurring terms.
+interested to analyse frequently occurring terms. DfR offers different grades of
+tokenization: n-grams for 1-3 words, e.g. unigrams, bigrams and trigrams. In
+case you are unfamiliar with the analysis of tokenized text, you could read
+the first few paragraphs of chapters 1 and 4 in https://www.tidytextmining.com
+as an introduction.
 
-Our analysis starts at the main page of DfR. We create a dataset by searching
-for
+Our analysis starts at the main page of [DfR](http://www.jstor.org/dfr/).
+We create a dataset^[An introduction on how to create datasets can be found
+on the page of DfR: http://www.jstor.org/dfr/about/creating-datasets]
+by searching for
 "inequality" and selecting "sociology" as our subject. To trim down the number
 of articles, we only select articles from 1997 to 2017. After
 logging in/creating an account, we select unigrams and bigrams. After unzipping
 the archives to a convenient location, we start our analysis.
-(#1)
+
 
 Up-front, we need to load some packages. `jstor` is currently not available from
 CRAN, but can be installed via `devtools`.
@@ -71,10 +77,9 @@ jstor_import(meta_files, out_file = "imported_metadata", .f = find_article,
 
 Since `jstor_import` writes the results to disk, we need to read the metadata
 from the newly created file.^[When reading this data without `guess_max = 2000`,
-a warning is raised because the column type was not recognized properly for one
-column. Increasing `guess_max` 
-helped in this instance.]
-(#2)
+a warning is raised because the column type is not being recognized properly for
+one column. Setting `guess_max` to `2000` helps to read the data correctly in
+this case.]
 
 ```{r, message=FALSE}
 imported_metadata <- read_csv("imported_metadata-1.csv", guess_max = 2000)
@@ -84,10 +89,8 @@ imported_metadata
 
 
 # Exploration
-Before diving into the analysis of ngrams, we might wish to take an explorative
-look at our metadata.
-
-The first thing to look at are the types of articles.
+Before diving into the analysis of n-grams, we might wish to take an explorative
+look at our metadata. The first thing to look at are the types of articles.
 ```{r}
 ggplot(imported_metadata, aes(article_type)) +
   geom_bar() +
@@ -125,12 +128,10 @@ research_articles <- imported_metadata %>%
   filter(!str_detect(article_title, "^Book Review"))
 ```
 
-## Filtering articles
-E.g.: It can happen that for one specific filter, one has to search in more than one column (captain obvious?)
 
 ## The moving wall - filtering articles by time
 Since JSTOR has a [moving wall](https://support.jstor.org/hc/en-us/articles/115004879547-JSTOR-s-Moving-Wall-Archive-vs-Current-Definitions),
-we should take a look at the number of articles per year in our dataset. (#3)
+we should take a look at the number of articles per year in our dataset.
 ```{r}
 research_articles %>% 
   ggplot(aes(pub_year)) +
@@ -140,7 +141,7 @@ research_articles %>%
 From this graph we can see an increase in research articles until 2010, after
 which the number of articles
 first tapers off, and then drops off sharply. For this reason we should
-exclude articles
+exclude articles at least
[…]
