### 2018-02-21 #189 @jsonbecker
Personal preference here-- in the ToC the Introduction vignette, the use of backticks to style the section headers resulted in links that don't look like links. I think that my preference would be to not have the fixed-width fonts for the functions in the headers such that the ToC is more clearly displaced as links.

When reading the description of `find_article`, `find_authors`, etc I don't have confidence about which should return more than one response. I think based on the pluralization, the expectation is that the XML always contains one article, which may have multiple authors, multiple references, and multiple footnotes. This is clearer in reading all of the examples, but if so, I think it'd be worth specifying whether the base XML will always have one article or more than one article up front. Additionally, I find the `kable` output to be appealing in the vignette and the swap from `find_article` and `find_authors` which use that output to just printing tables in the output is a little jarring for me. I'd probaby just use `kable` in all the examples.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

- [O] **Functionality:** Any functional claims of the software been confirmed.

    The functionality works with the supplied examples. Because the data set is not freely available though, it's hard to confirm with an authentic case. I have relatively high confidence this will not be a challenge based on the structured nature of the data set and clear code here.

- [O] **Performance:** Any performance claims of the software been confirmed.

    The functions work well and as described with the minimal data examples. There was some note about a desire for a technical review for speeding things up. I might suggest that if such a task was undertaken, it may make more sense to explore whether loading data into a sqlite database might be preferable to working with many XML files. It wouold be a pretty different package, but I can imagine a scenario where the primary goal of this package is loading the data into sqlite and then some convenience functions to access that data rather than from XML straight to tibbles. In my opinion, the fractured nature of the many XML files and atomiticity of the contents of those files suggest that I would want this data in a database for repeated access and not have to crawl the file system. I wonder if the data is too large for sqlite though under realistic loads. In that case, it would be more difficult to justify a package around the database load process, although I'd probably still pursue that if I were working with this package with my own database.

   The test coverage is solid and the tests are clear. The author may want to consider doing more tests that look at whether the resulting `tibble` is as expected as opposed to the more extensive element-level checks in some tests, e.g. more like `test-author-import.R` and less like `test-article.R`.

2 -2.5hrs

---

- I am not sure if I would name the first column response in `find_*` functions `basename`. Since this is intendended to be a file path or file name, I think it should be named `file_name` or similar to be more desciptive.
- I am not sure that I would be using the `basename_id` as the primary matching agent. I guess since articles do not consistently have a `doi`, `pub_id`, and `jcode`, you are trying to solve the problem by relating back to the file name. In my opinion, if the `article_doi`, `article_pub_id`, and `article_jcode` are a set where almost always only one value exists, then I would probably structure my data to have an `article_unique_id` and `article_id_type` or similar, based on gathering those three values. I'd far rather match on an externally valid identifier across files and other data than rely on my file path to the data. In that scenario, I'd think the file path is then an optional field really.
- I find it a little surprising that books don't contain some of the data available for articles, like footnotes and references. I somewhat expected to see equivalent functions across books and articles. I wonder if it would be valuable to have `find_authors` work for both books and articles. You already have functions to check if something is an `article` or a `book`, so why not have `find_authors` with a parameter for book or article. You can check for its presence, and if it's not user supplied, use your `validate_*` helpers to determine the type, then for `book` you can call `find_chapters(authors = TRUE)` internally and for `article` continue on the existing code path? Combined with my feedback above about the article identifiers, you could then return the same `tibble` structure for books or articles when calling `find_authors`, which seems convenient to me. This would also potentially avoid the `unnest` call by adding the `author_number` column to books like with articles.
- I see now by reading through the analyzing n-gram the motivation for having the file path in use for grabbing the redduced ngram2 text from this data set. While I think this motivates making file path an available piece of information, I still think my earlier comment about not using it as the core identifier stands to some degree. I did find this vignette to be helpful as someone not familiar with this data set.
- Given the file directory structure that comes from DfR requests, does it make sense to mimic that for test data as well? Are there support functions or processes used with the ngram2 files that you would find helpful to add as additional functions with this package?
- What parts of this data may make sense for loading into a sqlite database? The choice to go from XML to CSV is a great lowest denominator, but I wonder if there isn't greater advantage to formalizing the data in sqlite. Maintaining a large number of files in a specific directory structure, even after transitioning from XML to CSV poses some concer to me, especially with the file name being a major identifier for joins.
- If CSV is the appropriate target for some of this work, I wonder if it may be worthwhile to specify classes prior to `write_csv` to have the best likelihood of reloading that data and not losing type info in the serialization/deserialization process. Similarly, should their be import functions that can pull the CSV output of `jstor_import` into R with some assertions/checks?
