### 2018-02-21 #189 @tklebel
Thanks @jsonbecker for the thorough review! I'm eager to respond to the more substantial questions (`csv` vs `sqlite` and `file_name` vs `other_identifier`), but I think I will wait for the review by @EccRiley, so I can address all comments in one go.
