### 2018-02-21 #189 @jsonbecker
Absolutely. I plan to add a few more overall thoughts but with deadline looming I really wanted to get it in! I realize my review heavily focused on a suggestion that may have limited value and is a bit of my own preferences, but that’s largely because everything is otherwise in great shape!

http://www.jsonbecker.com/contact/

> On Feb 21, 2018, at 3:53 AM, Thomas Klebel <notifications@github.com> wrote:
>
> Thanks @jsonbecker for the thorough review! I'm eager to respond to the more substantial questions (csv vs sqlite and file_name vs other_identifier), but I think I will wait for the review by @EccRiley, so I can address all comments in one go.
>
