### 2018-02-26 baf1885
re-knit README with new pandoc/RStudio
