+++ README.md
@@ -1,30 +1,45 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-jstor
-=====
 
-[![Travis build status](https://travis-ci.org/tklebel/jstor.svg?branch=master)](https://travis-ci.org/tklebel/jstor) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/tklebel/jstor?branch=master&svg=true)](https://ci.appveyor.com/project/tklebel/jstor) [![Coverage status](https://codecov.io/gh/tklebel/jstor/branch/master/graph/badge.svg)](https://codecov.io/github/tklebel/jstor?branch=master) [![CRAN status](http://www.r-pkg.org/badges/version/jstor)](https://cran.r-project.org/package=jstor) [![CRAN\_Download\_Badge](http://cranlogs.r-pkg.org/badges/jstor)](https://CRAN.R-project.org/package=jstor) [![rOpenSci badge](https://badges.ropensci.org/189_status.svg)](https://github.com/ropensci/onboarding/issues/189) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1169862.svg)](https://doi.org/10.5281/zenodo.1169862)
+# jstor
 
-The tool [Data for Research (DfR)](http://www.jstor.org/dfr/) by JSTOR is a valuable source for citation analysis and text mining. `jstor` provides functions and suggests workflows for importing datasets from DfR. It was developed to deal with very large datasets which require an agreement, but can be used with smaller ones as well.
+[![Travis build
+status](https://travis-ci.org/tklebel/jstor.svg?branch=master)](https://travis-ci.org/tklebel/jstor)
+[![AppVeyor Build
+Status](https://ci.appveyor.com/api/projects/status/github/tklebel/jstor?branch=master&svg=true)](https://ci.appveyor.com/project/tklebel/jstor)
+[![Coverage
+status](https://codecov.io/gh/tklebel/jstor/branch/master/graph/badge.svg)](https://codecov.io/github/tklebel/jstor?branch=master)
+[![CRAN
+status](http://www.r-pkg.org/badges/version/jstor)](https://cran.r-project.org/package=jstor)
+[![CRAN\_Download\_Badge](http://cranlogs.r-pkg.org/badges/jstor)](https://CRAN.R-project.org/package=jstor)
+[![rOpenSci
+badge](https://badges.ropensci.org/189_status.svg)](https://github.com/ropensci/onboarding/issues/189)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1169862.svg)](https://doi.org/10.5281/zenodo.1169862)
+
+The tool [Data for Research (DfR)](http://www.jstor.org/dfr/) by JSTOR
+is a valuable source for citation analysis and text mining. `jstor`
+provides functions and suggests workflows for importing datasets from
+DfR. It was developed to deal with very large datasets which require an
+agreement, but can be used with smaller ones as well.
 
 The most important set of functions is a group of `find_*` functions:
 
--   `find_article`
--   `find_authors`
--   `find_references`
--   `find_footnotes`
--   `find_fulltext`
--   `find_book`
--   `find_chapters`
+  - `find_article`
+  - `find_authors`
+  - `find_references`
+  - `find_footnotes`
+  - `find_fulltext`
+  - `find_book`
+  - `find_chapters`
 
-All functions which are concerned with meta data (therefore excluding `find_fulltext`) operate along the same lines:
+All functions which are concerned with meta data (therefore excluding
+`find_fulltext`) operate along the same lines:
 
 1.  The file is read with `xml2::read_xml()`.
 2.  Content of the file is extracted via XPATH or CSS-expressions.
[…]
