### 2018-02-26 89f0acf
ad lifecycle badge
