### 2018-03-12 f5a42e6
improve documentation regarding endnotes
closes #43
