+++ NEWS.md
@@ -1,3 +1,5 @@
+# jstor 0.2.6.9000
+
 # jstor 0.2.6
 
 * added lengthy case study at https://tklebel.github.io/jstor/articles/analysing-n-grams.html
