+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: jstor
 Title: Read Data from JSTOR/DfR
-Version: 0.2.6
+Version: 0.2.6.9000
 Authors@R: person("Thomas", "Klebel", email = "thomas.klebel@uni-graz.at", 
                   role = c("aut", "cre"))
 Description: Functions and helpers to import metadata and full-texts delivered
@@ -9,7 +9,7 @@ Depends: R (>= 3.1)
 License: GPL-3 | file LICENSE
 Encoding: UTF-8
 LazyData: true
-Imports: 
+Imports:
     dplyr,
     purrr,
     xml2,
