### 2018-03-12 3c0b97f
update NEWS
