### 2018-03-12 tklebel/jstor#44 @tklebel
Address rOpenSci review requests
Review at: https://github.com/ropensci/onboarding/issues/189

# Review by jsonbecker
- [x] change `basename_id` to `file_name`.
- [x] explain why file_name is a good identifier (seconded by elinw)
- [x] think about export option to sqlite
- [x] add import function to read csvs from `jstor_import` #40
- [x] clarify, whether `find_authors` works on books or not
- [x] check, if `find_footnotes` and `find_references` could be used for books too
- [x] fix vignette formatting
- [x] improve error messages for find_references to say, that references are not available for books

# Review by elinw
- [x] add vignette with "known quirks" and how to handle them
- [ ] check aspect with repeated endnotes
