### 2018-03-12 a01de0c
do not assume that a missing lpage means the article only has one page
