### 2018-03-22 6eca551
address comments from Matthias for n-gram vignette
closes #46
