### 2018-04-05 #189 @tklebel
@karthik no worries. @elinw seems to have been trying out the package, maybe she would be willing to do a review.
