### 2018-04-30 1d79e5d
allow specification of either n_batches or files_per_batch
closes #51
