### 2018-05-04 019698f
rework read_jstor to allow for parts of zip files to be read
#1
