+++ tests/testthat/test-helpers.R
@@ -1,6 +1,7 @@
 context("test-helpers.R")
 
 invalid_file <- "testfiles/invalid_file.xml"
+standard_case <- "testfiles/standard_case.xml"
 
 # tests ------
 test_that("Files other than `book` and `article` raise an error", {
@@ -12,5 +13,23 @@ test_that("Files other than `book` and `article` raise an error", {
 
 test_that("Warnings for invalid URI are suppressed", {
   expect_silent(read_jstor(invalid_file))
+})
+
+
+test_that("read_jstor works", {
+  expect_equal(read_jstor(standard_case), xml2::read_xml(standard_case))
+  expect_error(read_jstor("test.txt"))
+  
+  zip_loc <- specify_zip_loc("testfiles/standard_case.zip", 
+                             "standard_case.xml")
+  expect_equal(read_jstor(zip_loc), xml2::read_xml(standard_case))
+  
+})
+
+
+test_that("Construction of zip location works", {
+  zip_loc <- specify_zip_loc("testfiles/standard_case.zip", 
+                             "standard_case.xml")
+  expect_s3_class(zip_loc, "jstor_zip")
 })
 
