### 2018-05-04 7c8d022
acknowledge code parts from readr
