### 2018-05-07 e60d39e
improve docs with example
