### 2018-05-07 2532996
check if there is any valid output to write
