### 2018-05-07 #189 @elinw
**This is still a draft**

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

JSTOR is a really important data source for the study of scholarly publication on issues.  JSTOR as a data source has wide coverage over many disciplines and also includes sources going back in time.  As a data source JSTOR has some challenges and features that are a bit different than others.  First,
there is not an API in the form of a REST interface. Instead, researchers fill out a web form and request the results of particular queries.  Then they are emailed a zip file containing potentially thousands (or more) of individual files including the xml data for the item and the n-gram data for that item in separate files for unigrams, bigrams and trigrams (assuming all of these were requested). These are organized into a
deeply nested file structure that ends with a folder for each item (article or book).  Obtaining full text data is a
separate process involving a user agreement. While not fundamentally  different than working with a web API that returns xml, it definitely feels different to the end user.

All of this to say that it is extremely helpful to have a package to help with managing this. Specifically the `jstor` package gets the data read in from all of the files into csv files separated by the type of content
(articles, authors, references, footnotes). It then writes to CSV files. I think this package will encourage people to use this data source more.

The package itself is solid, test coverage is good, documentation is complete, and good practices are followed. It does exactly what it says and works as described. The code seemed fine, and in particular the ways that it it made batch size and number of cores easy to specify was helpful to me.  I appreciated the single interface to all the data types.

The README is helpful in defining the structure and consistency of the data.

## Comments

These are some additional comments for consideration.

I have a JSTOR basic file (not full text) from my research that I used in my testing.  It contains a bit over 1678 items,  42051 references, 12376335 bigrams of raw data.   I did my testing mainly on RStudioServer (open source) installed on CentOS with 4 cores, although I also did some on my macbook air.   I will say that I was happy to have a package that installed easily.

At first I was surprised that the data were read to csv files. However, after some thought I feel it is very useful to have this as the main outcome as it is then up to the user to decide what kind of storage
to use for further processing. I found it useful to have the stored files to return back to.  One thing I will mention is that I did find that for my data I could not use `read_csv()` without generating warnings, but
`read.csv()` worked fine. It might be helpful to include a discussion of storage options for post-processing, but I also don't think this package requires documentation of all aspects of a text analysis. What is more
important is to focus on preparing the data for clean importing into other text analysis packages.

 I say that, but I followed the instructions in the bigrams/tidy text vignette pretty exactly and found them incredibly helpful.  So I won't discourage doing more of this as you go forward with maintaining the package, especially for known quirks of JSTOR (see below).   I also appreciated that there was a focus on getting a simple set of descriptive graphs.

The vignettes are helpful and the new one in particular is a nice overview of the process of cleaning and analyzing the ngram data.   However, I think it would be stronger if it discussed some of the issues
specific to the JSTOR data. For example, in working with the package I ended up creating a list of what I think of as my JSTOR stop words because academic writing is a specialized language.

For example for the references data I have:
```
stop_ref1 <- c("", "References", "Notes", "Note", "Ibid.", "Bibliography",
               "Loc. cit.", "Id.", "Foonotes", "Ibid.", "Endnotes",
               "[Bibliography]", "Notes and References", "Works Cited",
               "Footnotes","Ibid", "[References]", "Statutes Cited",
               "Literatur", "Case Cited", "Footnote", "Ibid",
               "Cases Cited", "[Endnotes]", "References", "Reference"
               )

```
And for the bigrams I had this.

```
custom_stop_words <- c("ibid", "supra", "pp", "eg", "alia", "hoc",
                       "e.g.")

```
(It is very easy to add custom stop words, I could figure out how just from following the vignette and realizing
I could add my vectors to those in Tidytext)

I also did not necessarily want to get rid of all numbers but one kind of troublesome number was e.g. 1998b.  So it would probably be easy to write a function for that, though I ended up just listing out the ones I found.

Another example, with JSTOR unless you specify otherwise, you get articles in multiple languages. The language names are not consistently coded  so I used `dplyr::filter(language == "en" | language == "eng") `.   There are various things like that which you might want to mention.

I do not have a problem with using the baseid as the main ID.  This potentially lets me work with other interfaces to JSTOR, is consistently produced over time and is unique.   If I have data from separate JSTOR queries it lets me find overlap easily.

---
