### 2018-05-07 8c61a68
handle zip files in get_basename
