### 2018-05-07 92e8ade
remove file path validation from import functions
(since it is incorporated now in read_jstor)
