+++ R/authors.R
@@ -21,8 +21,6 @@
 #' @examples 
 #' find_authors(jstor_example("sample_with_references.xml"))
 find_authors <- function(file_path) {
-  validate_file_path(file_path, "xml")
-
   xml_file <- read_jstor(file_path)
 
   if (identical(xml2::xml_name(xml_file), "article")) {
