+++ R/article.R
@@ -38,8 +38,6 @@
 #' @examples 
 #' find_article(jstor_example("sample_with_references.xml"))
 find_article <- function(file_path) {
-  validate_file_path(file_path, "xml")
-
   xml_file <- read_jstor(file_path)
 
   validate_article(xml_file)
