### 2018-05-07 d915176
check for length before attempting to read
