### 2018-05-08 #189 @elinw
Oh I meant to mention that the vignette might want to mention that installing igraph can be more complex than installing a normal package.
