### 2018-05-08 6ffe03f
check length properly
