### 2018-05-08 3a6eb1d
add test for multiple inputs
