### 2018-05-08 a281a9b
check for input length only if not "jstor_zip"
