### 2018-05-08 485dc65
first stab at import spec functions
