### 2018-05-08 c42e38c
first step to combine reading from zip and using an import_spec
currently batches don't work, and the name for import_spec might change
