### 2018-05-08 7c61b25
add function to preview contents of a jstor zip file
