+++ NAMESPACE
@@ -8,6 +8,7 @@ export(find_chapters)
 export(find_footnotes)
 export(find_references)
 export(get_basename)
+export(jst_preview_zip)
 export(jstor_example)
 export(jstor_import)
 export(read_full_text)
