### 2018-05-08 f2e93be
reorder outpaths according to initial order
