### 2018-05-08 #189 @elinw
Okay one more thing about jstor idiosyncrasies.   This is something I realized today (and relates to our prior discussion about end notes turning up in references).
 Reference lists/bibliographies have each source once, while endnotes may repeat the same source many times.  Therefore simply running a frequency on the sources will give you distorted results (I had a surprising article turn up as most cited, but it turned out that it was one article with end notes that cited it multiple times).  So I think that if you were willing, you could start a vignette that addresses some of these things and let users contribute strategies and code snippets for handling. Or you could write it yourself since you have a lot of experience, but I could see that getting a bit overwhelming unless you decide to write an article about this.
