+++ NAMESPACE
@@ -32,8 +32,13 @@ importFrom(readr,guess_encoding)
 importFrom(readr,locale)
 importFrom(readr,read_file)
 importFrom(readr,write_csv)
+importFrom(rlang,enquos)
+importFrom(rlang,eval_tidy)
+importFrom(rlang,get_expr)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_replace)
+importFrom(stringr,str_replace_all)
+importFrom(stringr,str_split)
 importFrom(tibble,new_tibble)
 importFrom(xml2,xml_child)
 importFrom(xml2,xml_children)
