### 2018-05-09 f6565f7
document jst_preview_zip()
