+++ man/jst_preview_zip.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/zip.R
+\name{jst_preview_zip}
+\alias{jst_preview_zip}
+\title{Preview contet of zip files}
+\usage{
+jst_preview_zip(zip_archive)
+}
+\arguments{
+\item{zip_archive}{A path to a .zip-file from DfR}
+}
+\value{
+The function returns a tibble with three columns:
+\itemize{
+\item \emph{type}: metadata or some form of ngram
+\item \emph{meta_type}: which type of metadata (book_chapter, journal article, ...)
+\item \emph{n}: a count for each category
+}
+}
+\description{
+This function gives you a quick preview about what a .zip-file from DfR
+contains.
+}
