### 2018-05-09 bb70917
check input to jst_import_spec
