### 2018-05-09 c3bdf27
clean output names if input was a list
