### 2018-05-09 55ae1ac
begin documenting functions
