+++ man/jst_define_import.Rd
@@ -1,0 +1,48 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/import_spec.R
+\name{jst_define_import}
+\alias{jst_define_import}
+\title{Define an import specification}
+\usage{
+jst_define_import(...)
+}
+\arguments{
+\item{...}{Named arguments with bare function names.}
+}
+\value{
+A specification of imports which is necessary for
+\code{\link[=jstor_import_zip]{jstor_import_zip()}}.
+}
+\description{
+Define which parts of a zip file should be converted via which functions.
+}
+\details{
+The function accepts the following names: article, book, report, pamphlet.
+The corresponding files from a .zip-archive will be imported via the supplied
+functions.
+}
+\examples{
+# articles will be imported via `find_article()` and `find_authors()`
+jst_define_import(article = c(find_article, find_author))
+
+# import all four types with one function each
+jst_define_import(article = find_article,
+                  book = find_book,
+                  report = find_book,
+                  pamphlet = find_article)
+                  
+# import all four types with multiple functions
+jst_define_import(article = c(find_article, find_authors, find_references),
+                  book = c(find_book, find_chapters),
+                  report = find_book,
+                  pamphlet = find_article)
+
+\dontrun{
+# define imports
+imports <- jst_define_import(article = c(find_article, find_author))
+
+# convert the files to .csv
+jstor_import_zip("my_archive.zip", out_file = "my_out_file", 
+                 import_spec = imports)
+}
+}
