+++ NAMESPACE
@@ -8,9 +8,11 @@ export(find_chapters)
 export(find_footnotes)
 export(find_references)
 export(get_basename)
+export(jst_define_import)
 export(jst_preview_zip)
 export(jstor_example)
 export(jstor_import)
+export(jstor_import_zip)
 export(read_full_text)
 importFrom(dplyr,as_data_frame)
 importFrom(dplyr,bind_rows)
