### 2018-05-09 1730e31
pass arguments show_progress and col_names along
