+++ NAMESPACE
@@ -34,6 +34,7 @@ importFrom(readr,guess_encoding)
 importFrom(readr,locale)
 importFrom(readr,read_file)
 importFrom(readr,write_csv)
+importFrom(rlang,abort)
 importFrom(rlang,enquos)
 importFrom(rlang,eval_tidy)
 importFrom(rlang,get_expr)
