### 2018-05-09 a611517
fix docs errors
