+++ R/batch_import_fun.R
@@ -168,7 +168,7 @@ jstor_convert_to_file <- function(in_paths, chunk_number, out_path, fun,
 #'
 #' # read from zip archive ------ 
 #' # define imports
-#' imports <- jst_define_import(article = c(find_article, find_author))
+#' imports <- jst_define_import(article = c(find_article, find_authors))
 #' 
 #' # convert the files to .csv
 #' jstor_import_zip("my_archive.zip", out_file = "my_out_file", 
