### 2018-05-09 c56c076
simplify argument checking by switching from flatten to unlist
