### 2018-05-09 8f89a87
stop if specified import is missing in zip-file
