### 2018-05-09 aa4349e
change argument order and rename function for jst_import_zip
