### 2018-05-09 f34ec63
adapt and test get_basename to work correctly
