### 2018-05-09 93597b4
add support for reading ngrams
