+++ man/jst_read_ngram.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ngram.R
+\name{jst_read_ngram}
+\alias{jst_read_ngram}
+\title{Read ngram data}
+\usage{
+jst_read_ngram(file)
+}
+\arguments{
+\item{file}{A path to a file.}
+}
+\value{
+A \link[tibble:tibble]{tibble::tibble} with two columns:
+\itemize{
+\item \emph{ngram}: the ngram term (unigram, bigram, trigram)
+\item \emph{n}: an integer for the number of times the term occured in the original
+file
+}
+}
+\description{
+Read in data on ngrams via \link[readr:read_tsv]{readr::read_tsv}.
+}
+\details{
+This function is mainly useful when it is used in together with
+\link{jst_import_zip}, where you can use it to specify reading in ngram data.
+}
