+++ man/jst_define_import.Rd
@@ -17,7 +17,8 @@ A specification of imports which is necessary for
 Define which parts of a zip file should be converted via which functions.
 }
 \details{
-The function accepts the following names: article, book, report, pamphlet.
+The function accepts the following names: article, book, report, pamphlet,
+ngram1, ngram2, ngram3.
 The corresponding files from a .zip-archive will be imported via the supplied
 functions.
 }
@@ -25,6 +26,11 @@ functions.
 # articles will be imported via `find_article()` and `find_authors()`
 jst_define_import(article = c(find_article, find_authors))
 
+# define a specification for importing article metadata and unigrams (ngram1)
+jst_define_import(article = find_article,
+                  ngram1 = jst_read_ngram)
+                  
+                  
 # import all four types with one function each
 jst_define_import(article = find_article,
                   book = find_book,
