+++ NAMESPACE
@@ -11,6 +11,7 @@ export(get_basename)
 export(jst_define_import)
 export(jst_import_zip)
 export(jst_preview_zip)
+export(jst_read_ngram)
 export(jstor_example)
 export(jstor_import)
 export(read_full_text)
