+++ tests/testthat/test-zip.R
@@ -1,4 +1,4 @@
-context("test-test-zip.R")
+context("test-zip.R")
 
 test_that("jst_preview_zip works", {
   correct_res <- tribble(
