### 2018-05-09 e56ae89
add tests for importing from .zip
