+++ tests/testthat/test-ngram.R
@@ -11,4 +11,8 @@ ngram2 <- tribble(~ngram, ~n,
 test_that("reading ngrams works", {
   expect_identical(jst_read_ngram("testfiles/ngram1.txt"), ngram1)
   expect_identical(jst_read_ngram("testfiles/ngram2.txt"), ngram2)
+  
+  expect_error(jst_read_ngram(c("path1.txt", "path2.txt")),
+               "file_path should be")
+  expect_error(jst_read_ngram("path.xml"), "Unknown input file")
 })
