### 2018-05-09 0b1a9a5
update news
closes #1
