### 2018-05-09 61cbc92
check if supplied rows fit actual data
