### 2018-05-15 cca702a
rework intro vignette with consistent output (knitr::kable)
