### 2018-05-15 2257994
don't use fixed width font for headings in vignette
