### 2018-05-15 fe70d5c
make error about references not being available more specific
