+++ R/helpers.R
@@ -24,9 +24,16 @@ validate_file_path <- function(file_path, type) {
 
 validate_article <- function(xml_file) {
   if (identical(xml2::xml_name(xml_file), "book")) {
-    stop(paste0("You are using `", sys.call(-1)[[1]], "` on a book. ",
-                "Please use `find_book` or `find_chapter` instead."),
-                call. = FALSE)
+    original_call <- as.character(sys.call(-1)[[1]])
+    if (any(original_call %in% c("find_references", "find_article"))) {
+      stop("You are using `", original_call, "` on a book. ",
+           "Neither footnotes nor references are available for books.",
+           call. = FALSE)
+    } else {
+      stop("You are using `", original_call, "` on a book. ",
+           "Please use `find_book` or `find_chapter` instead.",
+           call. = FALSE)
+    }
   } else if (!identical(xml2::xml_name(xml_file), "article")) {
     stop("Unknown input file.")
   }
