### 2018-05-15 #189 @tklebel
# Response to reviews
Thank you @jsonbecker and @elinw for your review and the helpful feedback! Although both of you mention, that your reviews are not finished, I will start working on your comments, so we can finish the onboarding hopefully until useR! (where I'll give a presentation about the package).

There are a couple requests for changes, and many thoughts for further consideration from you. I will outline my plan to address all issues in this post, and then work through them over the next two weeks.

From your feedback, I can distill three main areas of concern:

1. How data are exported after they have been brought into rectangular form, and how these data can later be re-imported.
2. The issue of using "basename_id" as the unique identifier.
3. Known quirks which are inherent to data from JSTOR/DfR and how to deal with them.

I will discuss each of the topics, and then proceed to the remaining comments.

## Data export and re-import
Depending on the size of the dataset under consideration, it might be worthwile putting the converted data into some sort of SQL-database. The package was actually designed to help with this. I separated the extraction functions in a way, so space on disk could be saved, because metadata would not have to be repeated for each reference the article has, etc. So the initial desing already planned for this structure in separate tables, which could be easily joined with dplyr.
As you both agree, exporting to `.csv` might be a good lowest denominator. However I would probably prefer not to implement the capability to write directly to sqlite in `jstor_import`, since I find it difficult to get the balance right about how much control users shoud have: if there is the option to write to sqlite, should there be options for creating new tables, appending, overwriting, etc.?

That being said, I have been busy last week at implementing a new function (`jst_import_zip`), which lets you import files directly from the zip-archive. Users can specify which parts they want to be converted via which functions (like `article = c(find_article, find_references), ngram1 = jst_read_ngram`). I hope, this function will make it a lot easier to deal with the multiple different sources which are available from JSTOR, and it removes the need to extract the zip (which saves quite
some disk space). The current output of this function are multiple csv-files with different file names refering to the used functions. I could very well imagine adding the option to write the output in a single sqlite-db. This would have the elegance of going from one zip-file directly to one .sqlite-file, without the hassle of any intermediate files (I don't know how I would solve the case where
you have multiple zip-archives, because the data were to large for one archive (they split up special requests into files of max 1.5GB)).
I would be happy about feedback on this idea! Since the reviews were initially for a version without `jst_import_zip`, I would prefer to keep the scope and consider `jst_import_zip` and related issues as an area for future improvements.

----

Regarding the further use of exported .csv-files, I could imagine two helper functions:

1. One to combine multiple output files into one file. When importing metadata or references for large datasets, it makes sense  to use batches, since you might not know, how big the resulting file will be. After conversion via `jstor_import`, you will most likely want to combine all batches into a single file, and this could be done via a helper function. As a name, I could imagine `jst_combine_outputs`. I would probaly opt for a "dumb" function which accepts paths as input, and simply imports, combines, and exports the content.
2. A simple helper function to import files which were exported from `jstor_import`. I could try to determine the type of content from the number, or, when available, name of columns: find_article hast 17 cols, find_book has 13, find_chapters has 9, find_authors has 6, and so forth. Other than the names of columns (if they are present in the .csv) or the number of columns I don't know how I would determine, what content the file has and how to read it. All this is only to avoid using `read_csv(., guess_max = 3000)` or similar, since the warnings mainly come from `read_csv` not guessing the column_types correctly.

## basename_id as unique identifier
I would agree with @elinw, that using the name of the files as unique identifier is ok ;) From my understanding, it is the only unique attribute of those files, which is the same for all articles. Articles may have up to three different types of ids besides the filename, and simply combining them might lead to unexpected issues. The filename is most likely consistent and allows to match metadata with ngrams/full-text-data. However, I like the suggestion for renaming it to `file_name`. This is more straightforward.

## Known quirks about the data
I agree with @elinw, that there are many quirks with data from DfR, some of which we know about, some of which we do not. I like the idea of having a vignette that collects known quirks and how to deal with them. There are currently a few helpers in the source which are not exported, like `find_total_pages`, which simply calculates the total number of pages from first and last page. This is helpful, but may lead to errors, when the page range for an article is something like: `50-70; 310`, which simply means, the body of the article is from 50-70, and there is a correction/erratum on page 310. Currently, calculating the total number of pages would result in 261 pages, since 50 is the first, and 310 the last page. There are many more quirks, and I cannot claim to have a good grasp over their majority. Especially, since I only worked with one very specific dataset, I might have found a few quirks in my data, but there might be other problems in other journals or for other applications. Therefore starting a vignette, and letting users contribute their knowledge (if they are willing to) sounds like a great idea to me.

## Issues with a simple answer

- find_authors works for books too.
- find_footnotes and find_references don't work for books, since I have not seen an example of a xml-file from a book containing
data about footnotes or references.

# Tasks to undertake
- [x] Improve formatting of vignette by using `knitr::kable()` and formatting headings in normal font.
- [ ] Clarify the the structure of data in intro vignette: xml-files only contain one article, one book (with multiple chapters), etc.
- [ ] Create helpers to re-import data after conversion with `jstor_import`.
- [ ] Write first version of vignette about known quirks.
- [x] Make error about references/footnotes not being available for books more specific.

# Other issues
## Multiple endnotes
@elinw could you point me to the article with multiple endnots so I can take a look?

## Function naming
I thought about changing names for functions across the package. After I posted the review, I had a break from the package for about 1 month. Coming back, I could not remember the names of the functions for a quick analysis :/ I'd like to at least add a prefix to all functions or change the current prefix. The tidyverse styleguide encourages the use of verbs, that's why I had `find_*`. Unfortunately the verb "find" might not be obvious in this context. A simple suggestion could be like the following:

- find_article -> jst_article
- find_authors -> jst_authors
- ...
- jstor_import -> jst_import
- jst_import_zip stays the same

I opted for "jst", since "js" might be similar to javascript, and the default for RStudio autocomplete is to display after three characters anyways.

What do you think about the proposed naming change?

-----

Please let me know if you agree with my roadmap or in case I overlooked an issue you raised.
