### 2018-05-15 5366dfc
extract page_range
closes #45
