### 2018-05-15 da90a8e
extract journal_title too
closes #41
