### 2018-05-15 9e9ff84
new function jst_get_journal_overview
closes #50
