+++ NAMESPACE
@@ -9,6 +9,7 @@ export(find_footnotes)
 export(find_references)
 export(get_basename)
 export(jst_define_import)
+export(jst_get_journal_overview)
 export(jst_import_zip)
 export(jst_preview_zip)
 export(jst_read_ngram)
