### 2018-05-15 8e69ee8
ungroup in jst_preview_zip and add example
