### 2018-05-16 e6478d0
handle new citation type and throw error if citation type is not known
closes #48
