### 2018-05-16 4b72929
expand intro vignette on file structure and number of rows in output
