### 2018-05-16 7e62028
showcase matching of article with references better
