### 2018-05-16 ceba372
remove backticks in vignette title
