### 2018-05-16 f122489
change layout and template for pkgdown
