### 2018-05-16 611a080
format number of files which are to be imported
closes #47
