### 2018-05-16 08195ea
don't convert pages to int by default
closes #38
