### 2018-05-16 d59c057
typos in docs
