### 2018-05-16 3b380bc
jst_read_ngram now adds basename_id
