+++ R/re-import.R
@@ -1,0 +1,229 @@
+jst_combine <- function(path, custom_name_part = NULL, overwrite = FALSE,
+                        clean_up = FALSE) {
+  
+  files <- list.files(path, pattern = "-\\d+.csv", full.names = T)
+  
+  splitted_paths <- tibble::tibble(files = files) %>% 
+    mutate(group = stringr::str_remove(files, "-\\d+\\.csv$")) %>% 
+    split(.$group) %>% 
+    purrr::map(~dplyr::pull(.data = ., files))
+  
+  out_names <- paste0(path, "/", custom_name_part, "combined_", 
+                      basename(names(splitted_paths)), ".csv")
+  
+  if (any(file.exists(out_names)) && !overwrite) {
+    abort(paste0("The file(s) `", paste0(out_names, collapse = "`, `"),
+                 "`` already exists. Do you want",
+                 " `overwrite = F`?"))
+  }
+  
+  
+  helper_fun <- function(x, path) {
+    message("Re-importing ", length(x), " batches.")
+    re_imported <- purrr::map_df(x, jst_re_import)
+    
+    message("Writing combined file `", path, "` to disk.")
+    write_csv(re_imported, path = path)
+  }
+  
+  purrr::walk2(splitted_paths, out_names, helper_fun)
+
+  if (clean_up) {
+    message("Deleting original batches.")
+    purrr::walk(splitted_paths, file.remove)
+  }
+  
+}
+
+
+jst_re_import <- function(file) {
+  sample_row <- readr::read_lines(file, n_max = 1) %>% 
+    stringr::str_split(",") %>% 
+    purrr::flatten_chr()
+  
+  # match by name
+  matches <- list(
+    names(article_cols$cols),
+    names(author_cols$cols),
+    names(book_cols$cols),
+    names(chapter_cols$cols),
+    names(chapter_w_authors$cols),
+    names(footnote_cols$cols),
+    names(reference_cols$cols),
+    names(ngram_cols$cols)
+  ) %>% 
+    purrr::map_lgl(identical, sample_row)
+  
+  
+  if (any(matches)) {
+    switch(which(matches),
+           article = read_csv(file, col_types = article_cols),
+           authors = read_csv(file, col_types = author_cols),
+           book = read_csv(file, col_types = book_cols),
+           chapter = read_csv(file, col_types = chapter_cols),
+           chapter_w_authors = read_csv(file, col_types = chapter_w_authors),
+           footnotes = read_csv(file, col_types = footnote_cols),
+           references = read_csv(file, col_types = reference_cols),
+           ngram = read_csv(file, col_types = ngram_cols))
+  } else {
+    # match by column length
+    matches <- c(
+      article = 19L,
+      article_old = 17L,
+      authors = 6L,
+      book = 13L,
+      chapter = 9L,
+      chapter_w_authors = 15L,
+      ngram = 3L
+    ) %>% 
+      purrr::map_lgl(identical, length(sample_row))
+    
+    if (any(matches)) {
+      switch(which(matches),
+             article = read_csv(file, col_types = article_cols,
+                                col_names = names(article_cols$cols)),
+             article_old = read_csv(file, col_types = article_cols_old,
+                                    col_names = names(article_cols_old$cols)),
+             authors = read_csv(file, col_types = author_cols,
+                                col_names = names(author_cols$cols)),
+             book = read_csv(file, col_types = book_cols,
+                             col_names = names(book_cols$cols)),
+             chapter = read_csv(file, col_types = chapter_cols,
+                                col_names = names(chapter_cols$cols)),
+             chapter_w_authors = read_csv(file, col_types = chapter_w_authors,
+                                          col_names = names(chapter_w_authors$cols)),
[…]
