+++ NAMESPACE
@@ -32,8 +32,12 @@ importFrom(purrr,map)
 importFrom(purrr,map_chr)
 importFrom(purrr,map_if)
 importFrom(purrr,map_lgl)
+importFrom(readr,col_character)
+importFrom(readr,col_integer)
+importFrom(readr,cols)
 importFrom(readr,guess_encoding)
 importFrom(readr,locale)
+importFrom(readr,read_csv)
 importFrom(readr,read_file)
 importFrom(readr,write_csv)
 importFrom(rlang,abort)
