### 2018-05-16 ba33616
resolve namespaces
