### 2018-05-16 f618277
further namespace fixes
