+++ R/imports.R
@@ -24,4 +24,5 @@ NULL
 # global variables
 utils::globalVariables(c("references", "footnotes", "encoding", "authors", "i",
                          "Name", "Length", "Date", "chunk_number", ".",
-                         "fun_names", "meta_type", "type", "full_coverage"))
+                         "fun_names", "meta_type", "type", "full_coverage",
+                         "title", "journal_id"))
