### 2018-05-16 67fb2b7
improve function for combining outputs
