### 2018-05-16 ca6fb2c
check file path and file type before attempting to re-read
