+++ R/re-import.R
@@ -45,11 +45,18 @@ jst_combine_outputs <- function(path, out_path = NULL, overwrite = FALSE,
 
 
 jst_re_import <- function(file) {
+  file <- check_path(file)
+  
+  if (!identical(tools::file_ext(file), "csv")) {
+    abort(paste("Only .csv-files which were generated by `jstor_import` or",
+                "`jst_import_zip` can be re-imported."))
+  }
+  
   sample_row <- readr::read_lines(file, n_max = 1) %>% 
     stringr::str_split(",") %>% 
     purrr::flatten_chr()
   
-  # match by name
+  # match by column names
   matches <- list(
     article = names(article_cols$cols),
     authors = names(author_cols$cols),
