### 2018-05-16 fc670fc
improve, export and document functions for combining outputs #40 #44
