+++ man/jst_combine_outputs.Rd
@@ -1,0 +1,67 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/re-import.R
+\name{jst_combine_outputs}
+\alias{jst_combine_outputs}
+\title{Combine outputs from converted files}
+\usage{
+jst_combine_outputs(path, write_to_file = TRUE, out_path = NULL,
+  overwrite = FALSE, clean_up = FALSE)
+}
+\arguments{
+\item{path}{A path to a directory, containing .csv-files from
+\code{\link[=jstor_import]{jstor_import()}} or \code{\link[=jst_import_zip]{jst_import_zip()}}.}
+
+\item{write_to_file}{Should to combined data be written to a file?}
+
+\item{out_path}{A directory where to write the combined files.}
+
+\item{overwrite}{Should files be overwritten?}
+
+\item{clean_up}{Do you want to remove the original batch files? Use with
+caution.}
+}
+\description{
+\code{jst_combine_outputs()} helps you to manage the multitude of files you might
+receive after running \code{\link[=jstor_import]{jstor_import()}} or \code{\link[=jst_import_zip]{jst_import_zip()}} with more than
+one batch.
+}
+\details{
+Splitting the output of \code{\link[=jstor_import]{jstor_import()}} or \code{\link[=jst_import_zip]{jst_import_zip()}} might be done
+for multiple reasons, but in the end you possibly want to combine all outputs
+into one file/data.frame. This function makes a few assumptions in order to
+combine files:
+\itemize{
+\item Files with similar names (except for trailing dashes with numbers) will be
+combined into one file.
+\item The names of the combined files are determined from the original files.
+If you want to combine \code{foo-1.csv} and \code{foo-2.csv}, the combined file will
+be \code{combined_foo.csv}.
+}
+}
+\examples{
+# set up a temporary directory
+tmp <- tempdir()
+
+# find multiple files
+file_list <- rep(jstor_example("sample_with_references.xml"), 2)
+
+# convert and write to file
+jstor_import(file_list, "article", out_path = tmp, .f = find_article,
+             n_batches = 2)
+             
+# combine outputs
+jst_combine_outputs(tmp)
+list.files(tmp, "csv")
+
+# Trying to combine the files again raises an error.
+jst_combine_outputs(tmp)
+jst_combine_outputs(tmp, overwrite = T)
+
+# we can remove the original files too
+jst_combine_outputs(tmp, overwrite = T, clean_up = TRUE)
+list.files(tmp, "csv")
+
+}
+\seealso{
+\code{\link[=jst_re_import]{jst_re_import()}}
+}
