+++ NAMESPACE
@@ -8,10 +8,12 @@ export(find_chapters)
 export(find_footnotes)
 export(find_references)
 export(get_basename)
+export(jst_combine_outputs)
 export(jst_define_import)
 export(jst_get_journal_overview)
 export(jst_import_zip)
 export(jst_preview_zip)
+export(jst_re_import)
 export(jst_read_ngram)
 export(jstor_example)
 export(jstor_import)
