### 2018-05-16 f1a6878
fix typos
