### 2018-05-16 8adb535
stop if there are no files to combine and improve dont run failing example
