### 2018-05-16 a0619e4
add new functions to pkgdown index
