### 2018-05-16 fb1853a
modify example and fix typo
