+++ man/jst_combine_outputs.Rd
@@ -47,7 +47,7 @@ file_list <- rep(jstor_example("sample_with_references.xml"), 2)
 
 # convert and write to file
 jstor_import(file_list, "article", out_path = tmp, .f = find_article,
-             n_batches = 2)
+             n_batches = 2, show_progress = FALSE)
              
 # combine outputs
 jst_combine_outputs(tmp)
