### 2018-05-16 b7dff3b
further clarification in docs
