### 2018-05-16 3477f90
update readme with two more supported elements
