+++ README.Rmd
@@ -95,7 +95,7 @@ The following table gives an overview, which elements are supported by `jstor`.
 |journal-id (type="publisher-id")  |x                |x                   |
 |journal-id (type="doi")           |                 |x                   |
 |issn                              |x                |                    |
-|journal-title                     |x                |                    |
+|journal-title                     |x                |x                   |
 |publisher-name                    |x                |                    |
 |article-id (type="doi")           |x                |x                   |
 |article-id (type="jstor")         |x                |x                   |
@@ -109,6 +109,7 @@ The following table gives an overview, which elements are supported by `jstor`.
 |pub-date                          |x                |x                   |
 |fpage                             |x                |x                   |
 |lpage                             |                 |x                   |
+|page-range                        |                 |x                   |
 |product                           |x                |                    |
 |self-uri                          |x                |                    |
 |kwd-group                         |x                |                    |
