### 2018-05-17 8c19836
Endnotes are in references, not in footnotes
