+++ man/jst_re_import.Rd
@@ -19,7 +19,7 @@ names which match the names from any of the \code{find_*} functions, the file
 is read with the corresponding specifications. If no column names are
 recognized, files are recognized based on the number of columns. Since both
 references and footnotes have only two columns, the first line is inspected
-for either \code{"Referenc...|Bilbio..."} or \code{"Footnote...|Endnote..."}.
+for either \code{"Referenc...|Bilbio...|Endnote..."} or \code{"Footnote..."}.
 In case there is still no match, the file is read with
 \code{\link[readr:read_csv]{readr::read_csv()}} with \code{guess_max = 5000} and a warning is raised.
 }
