### 2018-05-17 0e6113b
remove trailing -ngram\\d when reading ngrams
