+++ tests/testthat/test-ngram.R
@@ -9,8 +9,8 @@ ngram2 <- tribble(~basename_id, ~ngram, ~n,
                   "ngram2", "Uncommon word", 5L)
 
 test_that("reading ngrams works", {
-  expect_identical(jst_read_ngram("testfiles/ngram1.txt"), ngram1)
-  expect_identical(jst_read_ngram("testfiles/ngram2.txt"), ngram2)
+  expect_identical(jst_read_ngram("testfiles/ngram1-ngram1.txt"), ngram1)
+  expect_identical(jst_read_ngram("testfiles/ngram2-ngram2.txt"), ngram2)
   
   expect_error(jst_read_ngram(c("path1.txt", "path2.txt")),
                "file_path should be")
