### 2018-05-17 e97247b
more clear language in vignette
