### 2018-05-17 303997c
check for internet before attempting to download
