### 2018-05-17 1232167
add option to suppress messages when downloading latest journal info
