### 2018-05-17 56f1a17
add tests for journal list
