+++ R/imports.R
@@ -15,9 +15,9 @@ NULL
 #' @name %>%
 #' @rdname pipe
 #' @keywords internal
-#' @export
 #' @importFrom magrittr %>%
 #' @usage lhs \%>\% rhs
+#' @noRd
 NULL
 
 
