### 2018-05-17 a8dbec0
check for existing path when trying to re-import
