+++ tests/testthat/test-re-import.R
@@ -1,0 +1,233 @@
+context("test-re-import.R")
+
+# # prepare data
+# jst_import_zip("inst/extdata/pseudo_dfr.zip", import_spec = jst_define_import(article = c(find_article, 
+#     find_authors, find_references, find_footnotes), book = c(find_book, find_chapters), 
+#     ngram1 = jst_read_ngram), out_path = "tests/testthat/testfiles/re-import/", 
+#     out_file = "with_col", col_names = T)
+# 
+# jst_import_zip("inst/extdata/pseudo_dfr.zip", import_spec = jst_define_import(article = c(find_article, 
+#     find_authors, find_references, find_footnotes), book = c(find_book, find_chapters), 
+#     ngram1 = jst_read_ngram), out_path = "tests/testthat/testfiles/re-import/", 
+#     out_file = "wo_col", col_names = F)
+# 
+# # create two batches by simply copying.
+# files <- list.files("tests/testthat/testfiles/re-import", full.names = T)
+# new_names <- str_replace_all(files, "-1", "-2")
+# file.copy(files, new_names)
+
+
+### jst_re-import ------
+# set up correct objects
+# book ----
+book <- structure(list(book_id = "j.ctt24hdz7", basename_id = "book-chapter-standard_book", 
+    discipline = "Political Science", book_title = "The 2006 Military Takeover in Fiji", 
+    book_subtitle = "A Coup to End All Coups?", pub_day = 30L, pub_month = 4L, 
+    pub_year = 2009L, isbn = "9781921536502; 9781921536519", publisher_name = "ANU E Press", 
+    publisher_location = "Canberra", n_pages = NA_integer_, language = "eng"), 
+    row.names = c(NA, -1L), class = c("tbl_df", "tbl", "data.frame"))
+
+# chapter -----
+chapter <- structure(list(book_id = c("j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", 
+    "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", 
+    "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", 
+    "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", 
+    "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", 
+    "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", 
+    "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7", 
+    "j.ctt24hdz7", "j.ctt24hdz7", "j.ctt24hdz7"), basename_id = c("book-chapter-standard_book", 
+    "book-chapter-standard_book", "book-chapter-standard_book", "book-chapter-standard_book", 
+    "book-chapter-standard_book", "book-chapter-standard_book", "book-chapter-standard_book", 
+    "book-chapter-standard_book", "book-chapter-standard_book", "book-chapter-standard_book", 
+    "book-chapter-standard_book", "book-chapter-standard_book", "book-chapter-standard_book", 
+    "book-chapter-standard_book", "book-chapter-standard_book", "book-chapter-standard_book", 
[…]
