### 2018-05-17 885e1d2
add option to raise or not raise warnings when importing unrecognized files
