### 2018-05-17 69d08ec
expand tests on re-importing
