### 2018-05-17 95b521a
make jst_combine_outputs work with a vector of inputs
