### 2018-05-17 ece4c54
tweak docs
