### 2018-05-18 bf90787
allow anonymous functions as input to import spec #53
