### 2018-05-18 f101835
sanitize function names
closes #53
