### 2018-05-18 76732da
unnest before writing to file
closes #55
