### 2018-05-18 #189 @jsonbecker
I would be happy with the renaming to `jst_*`, but I might still prefer "verbing" to suggest the functionality, e.g. `jst_get_article`. I wouldn't hold up a 👍 for that. I am satisfied given more context with the move toward `file_name` and using that as an identifier. I guess I can't wish JSTOR toward having a better data model ;).

I'm also good with database being out of scope.
