### 2018-05-28 acf5e45
deprecate old functions
