+++ R/example.R
@@ -12,6 +12,8 @@
 #' jstor_example()
 #' jstor_example("sample_with_references.xml") 
 jstor_example <- function(path = NULL) {
+  base::.Deprecated(msg = paste("`jstor_example` has been deprecated.",
+                                "Please use `jst_example` instead."))
   if (is.null(path)) {
     dir(system.file("extdata", package = "jstor"))
   } else {
