+++ R/references.R
@@ -29,6 +29,8 @@
 #' @examples 
 #' find_references(jstor_example("sample_with_references.xml"))
 find_references <- function(file_path) {
+  base::.Deprecated(msg = paste("`find_references` has been deprecated.",
+                                "Please use `jst_get_references` instead."))
   xml_file <- read_jstor(file_path)
 
   validate_article(xml_file)
