+++ R/article.R
@@ -40,6 +40,8 @@
 #' @examples 
 #' find_article(jstor_example("sample_with_references.xml"))
 find_article <- function(file_path) {
+  base::.Deprecated(msg = paste("`find_article` has been deprecated.",
+                                 "Please use `jst_get_article` instead."))
   xml_file <- read_jstor(file_path)
 
   validate_article(xml_file)
