+++ R/full_text.R
@@ -19,6 +19,8 @@ get_encoding <- function(filename) {
 #'
 #' @export
 read_full_text <- function(filename) {
+  base::.Deprecated(msg = paste("`read_full_text` has been deprecated.",
+                                "Please use `jst_get_fulltext` instead."))
   validate_file_path(filename, "txt")
 
   id <- get_basename(filename)
