### 2018-05-28 c4f259d
rename functions
