+++ R/article.R
@@ -1,6 +1,6 @@
 #' Extract meta information for articles
 #'
-#' `find_article()` extracts meta-data from JSTOR-XML files for journal
+#' `jst_get_article()` extracts meta-data from JSTOR-XML files for journal
 #' articles.
 #'
 #' @param file_path A `.xml`-file for a journal-article.
@@ -38,30 +38,28 @@
 #'
 #' @export
 #' @examples 
-#' find_article(jstor_example("sample_with_references.xml"))
-find_article <- function(file_path) {
-  base::.Deprecated(msg = paste("`find_article` has been deprecated.",
-                                 "Please use `jst_get_article` instead."))
+#' jst_get_article(jst_example("sample_with_references.xml"))
+jst_get_article <- function(file_path) {
   xml_file <- read_jstor(file_path)
-
+  
   validate_article(xml_file)
-
+  
   front <- xml_find_all(xml_file, "front")
   article <- xml_child(front, "article-meta")
-
+  
   # pages
   first_page <- extract_page(article, "fpage", convert = FALSE)
   last_page <- extract_page(article, "lpage", convert = FALSE)
-
+  
   basename_id <- list(basename_id = get_basename(file_path))
-
+  
   journal_ids <- extract_jcode(front)
   journal_title <- list(
     journal_title = extract_child(front, ".//journal-title")
   )
   
   article_ids <- extract_article_id(front)
-
+  
   out <- list(
     article_type = xml2::xml_attr(xml_file, "article-type"),
     article_title = extract_title(article),
@@ -77,7 +75,7 @@ find_article <- function(file_path) {
     last_page = last_page,
     page_range = extract_child(article, "page-range")
   )
-
+  
   dplyr::bind_cols(basename_id, journal_ids, journal_title, article_ids, out)
 }
 
