+++ R/footnotes.R
@@ -9,7 +9,7 @@
 #' data. Because of this, there might be other content present than footnotes.
 #' 
 #' In order to get all available information on citation data, you might need to
-#' combine `find_footnotes()` with `find_references()`.
+#' combine `jst_get_footnotes()` with `jst_get_references()`.
 #'
 #' @param file_path The path to the `.xml`-file from which footnotes should be
 #'   extracted.
@@ -20,10 +20,8 @@
 #' 
 #' @export
 #' @examples 
-#' find_footnotes(jstor_example("sample_with_footnotes.xml"))
-find_footnotes <- function(file_path) {
-  base::.Deprecated(msg = paste("`find_footnotes` has been deprecated.",
-                                "Please use `jst_get_footnotes` instead."))
+#' jst_get_footnotes(jstor_example("sample_with_footnotes.xml"))
+jst_get_footnotes <- function(file_path) {
   xml_file <- read_jstor(file_path)
 
   validate_article(xml_file)
