### 2018-05-28 b222e79
move deprecated functions to specific file
