+++ R/deprecated.R
@@ -1,0 +1,543 @@
+#' Extract meta information for articles
+#'
+#' `find_article()` extracts meta-data from JSTOR-XML files for journal
+#' articles.
+#'
+#' @param file_path A `.xml`-file for a journal-article.
+#'
+#' @return A `tibble` containing the extracted meta-data with the following
+#' columns:
+#' - basename_id *(chr)*: The filename of the original .xml-file. Can be used 
+#'   for joining with other parts (authors, references, footnotes, full-texts).
+#' - journal_doi *(chr)*: A registered identifier for the journal.
+#' - journal_jcode *(chr)*: A identifier for the journal like "amerjsoci" for
+#'   the "American Journal of Sociology".
+#' - journal_pub_id *(chr)*: Similar to journal_jcode. Most of the time either
+#'   one is present.
+#' - journal_title *(chr)*: The title of the journal.
+#' - article_doi *(chr)*: A registered unique identifier for the article.
+#' - article_jcode *(chr)*: A unique identifier for the article (not a DOI).
+#' - article_pub_id *(chr)*: Infrequent, either part of the DOI or the 
+#'   article_jcode.
+#' - article_type *(chr)*: The type of article (research-article, book-review,
+#'   etc.).
+#' - article_title *(chr)*: The title of the article.
+#' - volume *(chr)*: The volume the article was published in.
+#' - issue *(chr)*: The issue the article was published in.
+#' - language *(chr)*: The language of the article.
+#' - pub_day *(chr)*: Publication day, if specified.
+#' - pub_month *(chr)*: Publication month, if specified.
+#' - pub_year *(int)*: Year of publication.
+#' - first_page *(int)*: Page number for the first page of the article.
+#' - last_page *(int)*: Page number for the last page of the article.
+#' - page_range *(chr)*: The range of pages for the article.
+#'
+#' A note about publication dates: always the first entry is being extracted,
+#' which should correspond to the oldest date, in case there is more than one
+#' date.
+#'
+#' @export
+#' @examples 
+#' find_article(jstor_example("sample_with_references.xml"))
+find_article <- function(file_path) {
+  base::.Deprecated(msg = paste("`find_article` has been deprecated.",
+                                "Please use `jst_get_article` instead."))
+  xml_file <- read_jstor(file_path)
+  
+  validate_article(xml_file)
+  
+  front <- xml_find_all(xml_file, "front")
+  article <- xml_child(front, "article-meta")
+  
+  # pages
+  first_page <- extract_page(article, "fpage", convert = FALSE)
+  last_page <- extract_page(article, "lpage", convert = FALSE)
+  
+  basename_id <- list(basename_id = get_basename(file_path))
+  
+  journal_ids <- extract_jcode(front)
+  journal_title <- list(
+    journal_title = extract_child(front, ".//journal-title")
+  )
+  
+  article_ids <- extract_article_id(front)
+  
+  out <- list(
+    article_type = xml2::xml_attr(xml_file, "article-type"),
+    article_title = extract_title(article),
+    volume = extract_child(article, "volume"),
+    issue = extract_child(article, "issue"),
+    language = extract_child(article, ".//meta-value"),
+    # the XPATH for the dates grabs always the first date by default.
+    # dates like "May" get turned into NA
+    pub_day = extract_child(article, ".//day"),
+    pub_month = extract_child(article, ".//month"),
+    pub_year = extract_child(article, ".//year") %>% as.integer(),
+    first_page = first_page,
+    last_page = last_page,
+    page_range = extract_child(article, "page-range")
+  )
+  
[…]
