+++ man/jst_import.Rd
@@ -1,28 +1,25 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/batch_import_fun.R, R/deprecated.R
-\name{jst_import_zip}
-\alias{jst_import_zip}
-\alias{jstor_import}
+% Please edit documentation in R/batch_import_fun.R
+\name{jst_import}
+\alias{jst_import}
 \title{Wrapper for file import}
 \usage{
-jst_import_zip(zip_archive, import_spec, out_file, out_path = NULL,
-  col_names = TRUE, n_batches = NULL, files_per_batch = NULL,
-  cores = getOption("mc.cores", 1L), show_progress = TRUE, rows = NULL)
-
-jstor_import(in_paths, out_file, out_path = NULL, .f, col_names = TRUE,
+jst_import(in_paths, out_file, out_path = NULL, .f, col_names = TRUE,
   n_batches = NULL, files_per_batch = NULL, cores = getOption("mc.cores",
   1L), show_progress = TRUE)
 }
 \arguments{
-\item{zip_archive}{A path to a .zip-archive from DfR}
-
-\item{import_spec}{A specification from \link{jst_define_import}
-for which parts of a .zip-archive should be imported via which functions.}
+\item{in_paths}{A character vector to the \code{xml}-files which should be
+imported}
 
 \item{out_file}{Name of files to export to. Each batch gets appended by an
 increasing number.}
 
 \item{out_path}{Path to export files to (combined with filename).}
+
+\item{.f}{Function to use for import. Can be one of \code{jst_get_article},
+\code{jst_get_authors}, \code{jst_get_references}, \code{jst_get_footnotes}, \code{jst_get_book}
+or \code{jst_get_chapter}.}
 
 \item{col_names}{Should column names be written to file? Defaults to \code{TRUE}.}
 
@@ -36,22 +33,20 @@ n_batches, but not in conjunction.}
 \item{show_progress}{Displays a progress bar for each batch, if the session
 is interactive.}
 
+\item{zip_archive}{A path to a .zip-archive from DfR}
+
+\item{import_spec}{A specification from \link{jst_define_import}
+for which parts of a .zip-archive should be imported via which functions.}
+
 \item{rows}{Mainly used for testing, to decrease the number of files which
 are imported (i.e. 1:100).}
-
-\item{in_paths}{A character vector to the \code{xml}-files which should be
-imported}
-
-\item{.f}{Function to use for import. Can be one of \code{find_article},
-\code{find_authors}, \code{find_references}, \code{find_footnotes}, \code{find_book} or
-\code{find_chapter}.}
 }
 \value{
 Writes \code{.csv}-files to disk.
 }
 \description{
 This function applies an import function to a list of \code{xml}-files
-or a .zip-archive in case of \code{jstor_import_zip} and saves
+or a .zip-archive in case of \code{jst_import_zip} and saves
 the output in batches of \code{.csv}-files to disk.
 }
 \details{
@@ -95,13 +90,13 @@ progress bar is displayed for each batch.
 # find all files
 meta_files <- list.files(pattern = "xml", full.names = T)
 
-# import them via `find_article`
-jstor_import(meta_files, out_file = "imported_metadata", .f = find_article,
-             files_per_batch = 25000, cores = 4)
+# import them via `jst_get_article`
+jst_import(meta_files, out_file = "imported_metadata", .f = jst_get_article,
+           files_per_batch = 25000, cores = 4)
 
 # read from zip archive ------ 
 # define imports
-imports <- jst_define_import(article = c(find_article, find_authors))
+imports <- jst_define_import(article = c(jst_get_article, jst_get_authors))
 
 # convert the files to .csv
 jstor_import_zip("my_archive.zip", out_file = "my_out_file", 
