+++ man/jst_get_authors.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/deprecated.R
-\name{find_authors}
-\alias{find_authors}
+% Please edit documentation in R/authors.R
+\name{jst_get_authors}
+\alias{jst_get_authors}
 \title{Extract author information}
 \usage{
-find_authors(file_path)
+jst_get_authors(file_path)
 }
 \arguments{
 \item{file_path}{A \code{.xml}-file from JSTOR containing meta-data.}
@@ -14,7 +14,7 @@ A \code{tibble} containing the extracted authors. All empty fields are
 \code{NA_character}.
 }
 \description{
-\code{find_authors()} extracts information about authors from JSTOR-XML files.
+\code{jst_get_authors()} extracts information about authors from JSTOR-XML files.
 }
 \details{
 The function returns a \code{tibble} with the following six columns:
@@ -30,5 +30,5 @@ data.
 }
 }
 \examples{
-find_authors(jstor_example("sample_with_references.xml"))
+jst_get_authors(jst_example("sample_with_references.xml"))
 }
