+++ man/jst_get_filename.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/deprecated.R
-\name{get_basename}
-\alias{get_basename}
+% Please edit documentation in R/helpers.R
+\name{jst_get_filename}
+\alias{jst_get_filename}
 \title{Extract the basename of a path}
 \usage{
-get_basename(file_path)
+jst_get_filename(file_path)
 }
 \arguments{
 \item{file_path}{A path to a file}
