+++ man/jst_get_footnotes.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/deprecated.R
-\name{find_footnotes}
-\alias{find_footnotes}
+% Please edit documentation in R/footnotes.R
+\name{jst_get_footnotes}
+\alias{jst_get_footnotes}
 \title{Extract all footnotes}
 \usage{
-find_footnotes(file_path)
+jst_get_footnotes(file_path)
 }
 \arguments{
 \item{file_path}{The path to the \code{.xml}-file from which footnotes should be
@@ -26,8 +26,8 @@ therefore extracts the content of \code{fn-group} exactly as it appears in the
 data. Because of this, there might be other content present than footnotes.
 
 In order to get all available information on citation data, you might need to
-combine \code{find_footnotes()} with \code{find_references()}.
+combine \code{jst_get_footnotes()} with \code{jst_get_references()}.
 }
 \examples{
-find_footnotes(jstor_example("sample_with_footnotes.xml"))
+jst_get_footnotes(jstor_example("sample_with_footnotes.xml"))
 }
