+++ man/jst_get_book.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/deprecated.R
-\name{find_book}
-\alias{find_book}
+% Please edit documentation in R/books.R
+\name{jst_get_book}
+\alias{jst_get_book}
 \title{Extract meta information for books}
 \usage{
-find_book(file_path)
+jst_get_book(file_path)
 }
 \arguments{
 \item{file_path}{A \code{.xml}-file for a book or research report.}
@@ -36,8 +36,8 @@ which should correspond to the oldest date, in case there is more than one
 date.
 }
 \description{
-\code{find_book()} extracts meta-data from JSTOR-XML files for book chapters.
+\code{jst_get_book()} extracts meta-data from JSTOR-XML files for book chapters.
 }
 \examples{
-find_book(jstor_example("sample_book.xml"))
+jst_get_book(jst_example("sample_book.xml"))
 }
