+++ man/jst_read_fulltext.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/deprecated.R
-\name{read_full_text}
-\alias{read_full_text}
+% Please edit documentation in R/full_text.R
+\name{jst_read_fulltext}
+\alias{jst_read_fulltext}
 \title{Import full-text}
 \usage{
-read_full_text(filename)
+jst_read_fulltext(filename)
 }
 \arguments{
 \item{filename}{The path to the file.}

[…]
