+++ R/deprecated.R
@@ -485,7 +485,7 @@ read_full_text <- function(filename) {
 #' @export
 get_basename <- function(file_path) {
   base::.Deprecated(msg = paste("`get_basename` has been deprecated.",
-                                "Please use `jst_get_filename` instead."))
+                                "Please use `jst_get_file_name` instead."))
   if (inherits(file_path, "jstor_zip")) {
     file_path[["file_path"]] %>% 
       basename() %>% 
