+++ man/jst_get_chapters.Rd
@@ -20,7 +20,7 @@ columns:
 \itemize{
 \item book_id \emph{(chr)}: The book id of type "jstor", which is not a registered
 DOI.
-\item basename_id \emph{(chr)}: The filename of the original .xml-file. Can be used
+\item file_name \emph{(chr)}: The filename of the original .xml-file. Can be used
 for joining with other data for the same file.
 \item part_id \emph{(chr)}: The id of the part.
 \item part_label \emph{(chr)}: A label for the part, if specified.
