+++ man/jst_get_file_name.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/helpers.R
-\name{jst_get_filename}
-\alias{jst_get_filename}
+\name{jst_get_file_name}
+\alias{jst_get_file_name}
 \title{Extract the basename of a path}
 \usage{
-jst_get_filename(file_path)
+jst_get_file_name(file_path)
 }
 \arguments{
 \item{file_path}{A path to a file}
