+++ NAMESPACE
@@ -14,7 +14,7 @@ export(jst_get_article)
 export(jst_get_authors)
 export(jst_get_book)
 export(jst_get_chapters)
-export(jst_get_filename)
+export(jst_get_file_name)
 export(jst_get_footnotes)
 export(jst_get_journal_overview)
 export(jst_get_references)
