### 2018-05-28 939047e
update tests
