+++ tests/testthat/test-author-import.R
@@ -6,53 +6,53 @@ context("author-import")
 
 # import files
 test_file_no_author <- "testfiles/erratum.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 # contrib group but no authors
 test_file_no_author2 <- "testfiles/no-authors.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 test_file_single_author <- "testfiles/standard_case.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 test_file_multiple_authors <- "testfiles/multiple-authors.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 test_file_author_string <- "testfiles/author-string.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 test_file_author_prefix <- "testfiles/author-prefix.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 test_file_author_suffix <- "testfiles/author-suffix.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 test_file_book <- "testfiles/standard_book.xml" %>%
-  find_authors()
+  jst_get_authors()
 
 # nolint start
 
 # expected output
 single_author <- tribble(
-  ~basename_id,              ~prefix,       ~given_name, ~surname,     ~string_name, ~suffix,        ~author_number,
+  ~file_name,              ~prefix,       ~given_name, ~surname,     ~string_name, ~suffix,        ~author_number,
   "standard_case", NA_character_, "N. L.",  "Bor", NA_character_, NA_character_,  1L
 )
 
 multiple_authors <- tribble(
-  ~basename_id,                 ~prefix,       ~given_name, ~surname,  ~string_name, ~suffix,        ~author_number,
+  ~file_name,                 ~prefix,       ~given_name, ~surname,  ~string_name, ~suffix,        ~author_number,
   "multiple-authors", NA_character_, "Louis",     "Kaplow",  NA_character_, NA_character_,  1L,
   "multiple-authors", NA_character_, "Steven",    "Shavell", NA_character_, NA_character_,  2L
 )
 
 multiple_given_names <- tribble(
-  ~basename_id,                    ~prefix,       ~given_name,    ~surname, ~string_name,  ~suffix,        ~author_number,
+  ~file_name,                    ~prefix,       ~given_name,    ~surname, ~string_name,  ~suffix,        ~author_number,
   "multiple-given-names", NA_character_, "Seung Ho",      "Park", NA_character_, NA_character_,  1L,
   "multiple-given-names", NA_character_, "Roger",         "Chen", NA_character_, NA_character_,  2L,
   "multiple-given-names", NA_character_, "Scott",    "Gallagher", NA_character_, NA_character_,  3L
 )
 
 author_string <- tribble(
-  ~basename_id,              ~prefix,      ~given_name,    ~surname,     ~string_name,               ~suffix,        ~author_number,
+  ~file_name,              ~prefix,      ~given_name,    ~surname,     ~string_name,               ~suffix,        ~author_number,
   "author-string", NA_character_, NA_character_, NA_character_, " Michèle de la Pradelle ", NA_character_,  1L,
   "author-string", NA_character_, NA_character_, NA_character_, "Emmanuelle Lallement",     NA_character_,  2L
 )
@@ -63,7 +63,7 @@ no_authors <- tribble(
 )
 
 book_authors <- tribble(
-  ~basename_id, ~prefix, ~given_name, ~surname, ~string_name, ~suffix, ~author_number,
+  ~file_name, ~prefix, ~given_name, ~surname, ~string_name, ~suffix, ~author_number,
[…]
