+++ tests/testthat/test-zip.R
@@ -17,9 +17,9 @@ test_that("jst_preview_zip works", {
 # # test converting from zip-file
 # # setup the correct data
 # jst_import_zip("tests/testthat/testfiles/pseudo_dfr.zip",
-#                import_spec = jst_define_import(article = c(find_article,
-#                                                            find_authors),
-#                                                book = find_book,
+#                import_spec = jst_define_import(article = c(jst_get_article,
+#                                                            jst_get_authors),
+#                                                book = jst_get_book,
 #                                                ngram1 = jst_read_ngram),
 #                out_file = "correct", out_path = "tests/testthat/testfiles/")
 
@@ -27,16 +27,16 @@ test_that("importing from zip works", {
   dir.create(test_dir <- file.path(tempdir(), "testdir"))
   
   jst_import_zip("testfiles/pseudo_dfr.zip",
-                 import_spec = jst_define_import(article = c(find_article, 
-                                                             find_authors),
-                                                 book = find_book,
+                 import_spec = jst_define_import(article = c(jst_get_article, 
+                                                             jst_get_authors),
+                                                 book = jst_get_book,
                                                  ngram1 = jst_read_ngram),
                  out_file = "correct", out_path = test_dir)
   
   
-  expected_files <- c("correct_book_chapter_find_book-1.csv",
-                      "correct_journal_article_find_article-1.csv",
-                      "correct_journal_article_find_authors-1.csv",
+  expected_files <- c("correct_book_chapter_jst_get_book-1.csv",
+                      "correct_journal_article_jst_get_article-1.csv",
+                      "correct_journal_article_jst_get_authors-1.csv",
                       "correct_ngram1_jst_read_ngram-1.csv")
   files <- list.files(test_dir)
   
@@ -55,8 +55,8 @@ test_that("importing from zip works", {
 test_that("too many arguments for batches throw error", {
   expect_error(jst_import_zip("testfiles/pseudo_dfr.zip", out_file = "meta_data",
                               import_spec = jst_define_import(
-                                article = find_article,
-                                report = find_book
+                                article = jst_get_article,
+                                report = jst_get_book
                               ), out_path = test_dir,
                               n_batches = 1),
                "not available in the .zip-file: research_report"
@@ -68,7 +68,7 @@ test_that("wrong row selection raises an error", {
   expect_error(jst_import_zip("testfiles/pseudo_dfr.zip",
                               out_file = "meta_data",
                               import_spec = jst_define_import(
-                                article = find_article),
+                                article = jst_get_article),
                               rows = 10:20),
                "The selected rows do not exist within")
 })
