### 2018-05-28 011cd77
don't compute coverage for deprecated functions
