### 2018-05-28 9831feb
rename jst_read_fulltext -> jst_get_full_text
