+++ man/jst_get_full_text.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/full_text.R
-\name{jst_read_fulltext}
-\alias{jst_read_fulltext}
+\name{jst_get_full_text}
+\alias{jst_get_full_text}
 \title{Import full-text}
 \usage{
-jst_read_fulltext(filename)
+jst_get_full_text(filename)
 }
 \arguments{
 \item{filename}{The path to the file.}
