### 2018-05-28 184250e
use "get_" for ngrams too
