+++ man/jst_get_ngram.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/ngram.R
-\name{jst_read_ngram}
-\alias{jst_read_ngram}
+\name{jst_get_ngram}
+\alias{jst_get_ngram}
 \title{Read ngram data}
 \usage{
-jst_read_ngram(file)
+jst_get_ngram(file)
 }
 \arguments{
 \item{file}{A path to a file.}
