+++ NAMESPACE
@@ -16,14 +16,14 @@ export(jst_get_book)
 export(jst_get_chapters)
 export(jst_get_file_name)
 export(jst_get_footnotes)
+export(jst_get_full_text)
 export(jst_get_journal_overview)
+export(jst_get_ngram)
 export(jst_get_references)
 export(jst_import)
 export(jst_import_zip)
 export(jst_preview_zip)
 export(jst_re_import)
-export(jst_read_fulltext)
-export(jst_read_ngram)
 export(jstor_example)
 export(jstor_import)
 export(read_full_text)
