+++ man/jst_import.Rd
@@ -2,11 +2,16 @@
 % Please edit documentation in R/batch_import_fun.R
 \name{jst_import}
 \alias{jst_import}
+\alias{jst_import_zip}
 \title{Wrapper for file import}
 \usage{
 jst_import(in_paths, out_file, out_path = NULL, .f, col_names = TRUE,
   n_batches = NULL, files_per_batch = NULL, cores = getOption("mc.cores",
   1L), show_progress = TRUE)
+
+jst_import_zip(zip_archive, import_spec, out_file, out_path = NULL,
+  col_names = TRUE, n_batches = NULL, files_per_batch = NULL,
+  cores = getOption("mc.cores", 1L), show_progress = TRUE, rows = NULL)
 }
 \arguments{
 \item{in_paths}{A character vector to the \code{xml}-files which should be
