+++ man/jstor_import.Rd
@@ -1,28 +1,25 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/batch_import_fun.R, R/deprecated.R
-\name{jst_import_zip}
-\alias{jst_import_zip}
+% Please edit documentation in R/deprecated.R
+\name{jstor_import}
 \alias{jstor_import}
 \title{Wrapper for file import}
 \usage{
-jst_import_zip(zip_archive, import_spec, out_file, out_path = NULL,
-  col_names = TRUE, n_batches = NULL, files_per_batch = NULL,
-  cores = getOption("mc.cores", 1L), show_progress = TRUE, rows = NULL)
-
 jstor_import(in_paths, out_file, out_path = NULL, .f, col_names = TRUE,
   n_batches = NULL, files_per_batch = NULL, cores = getOption("mc.cores",
   1L), show_progress = TRUE)
 }
 \arguments{
-\item{zip_archive}{A path to a .zip-archive from DfR}
-
-\item{import_spec}{A specification from \link{jst_define_import}
-for which parts of a .zip-archive should be imported via which functions.}
+\item{in_paths}{A character vector to the \code{xml}-files which should be
+imported}
 
 \item{out_file}{Name of files to export to. Each batch gets appended by an
 increasing number.}
 
 \item{out_path}{Path to export files to (combined with filename).}
+
+\item{.f}{Function to use for import. Can be one of \code{find_article},
+\code{find_authors}, \code{find_references}, \code{find_footnotes}, \code{find_book} or
+\code{find_chapter}.}
 
 \item{col_names}{Should column names be written to file? Defaults to \code{TRUE}.}
 
@@ -36,15 +33,13 @@ n_batches, but not in conjunction.}
 \item{show_progress}{Displays a progress bar for each batch, if the session
 is interactive.}
 
+\item{zip_archive}{A path to a .zip-archive from DfR}
+
+\item{import_spec}{A specification from \link{jst_define_import}
+for which parts of a .zip-archive should be imported via which functions.}
+
 \item{rows}{Mainly used for testing, to decrease the number of files which
 are imported (i.e. 1:100).}
-
-\item{in_paths}{A character vector to the \code{xml}-files which should be
-imported}
-
-\item{.f}{Function to use for import. Can be one of \code{find_article},
-\code{find_authors}, \code{find_references}, \code{find_footnotes}, \code{find_book} or
-\code{find_chapter}.}
 }
 \value{
 Writes \code{.csv}-files to disk.
