### 2018-05-28 7936ece
update to jst_get_ngram properly
