+++ man/jst_define_import.Rd
@@ -28,7 +28,7 @@ jst_define_import(article = c(jst_get_article, jst_get_authors))
 
 # define a specification for importing article metadata and unigrams (ngram1)
 jst_define_import(article = jst_get_article,
-                  ngram1 = jst_read_ngram)
+                  ngram1 = jst_get_ngram)
                   
                   
 # import all four types with one function each
