### 2018-05-28 cae3498
avoid overly long paths and use jst_get_ngram
