### 2018-05-28 f006041
add title to yaml macros in vignette
