+++ vignettes/Introduction.Rmd
@@ -6,7 +6,7 @@ output:
   rmarkdown::html_vignette:
       toc: true
 vignette: >
-  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteIndexEntry{Introduction to jstor}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
