### 2018-05-28 d8f122d
no need to return anything after raising an error
