### 2018-05-28 60bb7fe
export helper for total pages
