+++ R/article.R
@@ -170,10 +170,3 @@ convert_page <- function(x) {
   }
 }
 
-find_total_pages <- function(first_page, last_page) {
-  dplyr::case_when(
-    !is.na(first_page) & is.na(last_page) ~ NA_real_,
-    !is.na(first_page) & !is.na(last_page) ~ last_page - first_page + 1,
-    TRUE ~ NA_real_
-  )
-}
