+++ NAMESPACE
@@ -20,6 +20,7 @@ export(jst_get_full_text)
 export(jst_get_journal_overview)
 export(jst_get_ngram)
 export(jst_get_references)
+export(jst_get_total_pages)
 export(jst_import)
 export(jst_import_zip)
 export(jst_preview_zip)
