### 2018-05-28 a97af74
start vignette about known quirks
