### 2018-05-28 5d280a3
update pkgdown reference list
