### 2018-05-28 85aee8b
Update NEWS.md
