### 2018-05-29 5058630
download file differently for windows
closes #60
