### 2018-05-29 c0a5c4e
new function jst_subset_ngrams
