+++ tests/testthat/test-ngram.R
@@ -16,3 +16,29 @@ test_that("reading ngrams works", {
                "file_path should be")
   expect_error(jst_get_ngram("path.xml"), "Unknown input file")
 })
+
+
+test_that("subsetting ngrams works", {
+  # this test is not very thorough
+  # should test other scenarios too, with multiple matching files etc.
+
+  
+  correct_out <- tribble(~file_name, ~ngram, ~n,
+                         "book-chapter-standard_book", "Common", 400L,
+                         "book-chapter-standard_book", "Uncommon", 5L)
+  
+  # create sample output
+  tmp <- tempdir()
+  jst_import_zip(jst_example("pseudo_dfr.zip"),
+                import_spec = jst_define_import(book = jst_get_book),
+                out_file = "test", out_path = tmp, show_progress = F)
+  
+  # re-import as our selection for which we would like to import ngrams
+  selection <- jst_re_import(paste0(tmp, "/test_book_chapter_jst_get_book-1.csv"))
+  
+  # get location of file
+  zip_loc <- jst_subset_ngrams(jst_example("pseudo_dfr.zip"), "ngram1", selection) 
+  
+  
+  expect_identical(jst_get_ngram(zip_loc[[1]]), correct_out)
+})
