+++ NAMESPACE
@@ -25,6 +25,7 @@ export(jst_import)
 export(jst_import_zip)
 export(jst_preview_zip)
 export(jst_re_import)
+export(jst_subset_ngrams)
 export(jstor_example)
 export(jstor_import)
 export(read_full_text)
