### 2018-05-29 59d0a68
new quirk about duplicated ngrams
