+++ vignettes/analysing-n-grams.Rmd
@@ -31,8 +31,8 @@ on the page of DfR: http://www.jstor.org/dfr/about/creating-datasets]
 by searching for
 "inequality" and selecting "sociology" as our subject. To trim down the number
 of articles, we only select articles from 1997 to 2017. After
-logging in/creating an account, we select unigrams and bigrams. After unzipping
-the archives to a convenient location, we start our analysis.
+logging in/creating an account, we select unigrams and bigrams. The resulting 
+`.zip`-file can be downloaded from the email which DfR sends.
 
 
 Up-front, we need to load some packages. `jstor` is currently not available from
@@ -51,41 +51,51 @@ library(visdat)
 theme_set(theme_bw())
 ```
 
-To import the files, we first need to locate them and generate an object with
-their corresponding paths. The following code assumes that you follow a workflow
+
+With the latest version of `jstor` we can now directly import files from a zip
+archive. We only need to specify, where the zip archive is located, which
+parts we want to extract, and where the resulting files should be saved. 
+
+Although we could import all ngrams at this point, this would be extremely 
+inefficient. It is thus best first to decide, which articles we want to analyze,
+and import the corresponding ngram-files afterwards.
+
+The following code assumes that you follow a workflow
 organised around projects within RStudio (refer to
 http://r4ds.had.co.nz/workflow-projects.html for further information). 
 
-```{r}
-# list files
-meta_files <- list.files(pattern = "xml", full.names = T, recursive = T)
-```
-
-Since we have a decent amount of articles, let's speed up the process of
-importing the metadata via parallel processing. `jstor_import` is a nice wrapper
-which takes care of setting up the processes and deals with any errors that
-might occur along the way.
-
 ```{r, eval=FALSE}
-jstor_import(meta_files, out_file = "imported_metadata", .f = find_article,
-             files_per_batch = 25000, cores = 4)
-#> Starting to import 23909 file(s).
+import_spec <- jst_define_import(article = jst_get_article)
+
+jst_import_zip("part1.zip", import_spec = import_spec, out_file = "part1",
+               cores = 4)
+jst_import_zip("part2.zip", import_spec = import_spec, out_file = "part2",
+               cores = 4)
+#> Processing files for journal_article with functions jst_get_article
 #> Processing chunk 1/1
 #>   |===================================================================| 100%
-#> Finished importing 23909 file(s) in 2.54 mins.
 ```
 
-Since `jstor_import` writes the results to disk, we need to read the metadata
-from the newly created file.^[When reading this data without `guess_max = 2000`,
-a warning is raised because the column type is not being recognized properly for
-one column. Setting `guess_max` to `2000` helps to read the data correctly in
-this case.]
+Since `jst_import_zip` writes the results to disk, we need to read the metadata
+from the newly created file. This is made easy by `jst_re_import` which ensures,
+that the data are read with the right column types.
 
-```{r, message=FALSE}
-imported_metadata <- read_csv("imported_metadata-1.csv", guess_max = 2000)
+```{r}
+imported_metadata <- c("part1_journal_article_jst_get_article-1.csv",
+                       "part2_journal_article_jst_get_article-1.csv") %>% 
+  map_df(jst_re_import)
 imported_metadata
 ```
 
+# Cleaning the data
+
+- convert years to integers
+- do other things from augment (unify journal id)
+
+```{r}
+imported_metadata <- imported_metadata %>% 
+  mutate(pub_year = as.integer(pub_year))
[…]
