### 2018-05-29 63afe1e
remove comment which lurked around
