+++ DESCRIPTION
@@ -10,19 +10,19 @@ License: GPL-3 | file LICENSE
 Encoding: UTF-8
 LazyData: true
 Imports:
-    dplyr,
-    tidyr,
-    purrr,
-    xml2,
+    dplyr (>= 0.7.4),
+    tidyr (>= 0.7.2),
+    purrr (>= 0.2.4),
+    xml2 (>= 1.2.0),
     magrittr,
-    stringr,
-    readr,
-    tibble,
-    rlang,
+    stringr (>= 1.3.0),
+    readr (>= 1.0.0),
+    tibble (>= 1.4.1),
+    rlang (>= 0.2.0),
     foreach,
     doSNOW,
     snow,
-    readxl,
+    readxl (>= 1.0.0),
     curl,
     pryr
 Suggests: testthat,
