### 2018-05-29 8bd22d3
replace snow and doSnow with furrr::future_map
closes #59
