+++ R/batch_import_fun.R
@@ -13,40 +13,19 @@
 #' @param col_names Should `col_names` be printed when exporting results? For
 #'   errors, col_names are always written to file.
 #' @param n_batches Total number of batches (for progress bar).
-#' @param cores Number of cores to use for parallel processing.
 #' @noRd
 jstor_convert_to_file <- function(in_paths, chunk_number, out_path, fun,
                                   col_names = FALSE, n_batches,
-                                  cores = getOption("mc.cores", 1L),
                                   show_progress = TRUE) {
   message("Processing chunk ", chunk_number, "/", n_batches)
 
   safe_fun <- purrr::safely(fun)
   
-  
-  # start cluster
-  cl <- snow::makeCluster(cores)
-  doSNOW::registerDoSNOW(cl)
-  
   # create progress bar if we are in console
-  if (interactive() && show_progress) {
-    pb <- utils::txtProgressBar(min = 0, max = length(in_paths), style = 3)
-    progress <- function(n) utils::setTxtProgressBar(pb, n)
-    opts <- list(progress = progress)
-  } else {
-    opts <- list()
-  }
-
+  progress <- ifelse(interactive() && show_progress, TRUE, FALSE)
   
   # compute results in parallel
-  parallel_result <- foreach::foreach(i = in_paths, .packages = "jstor",
-                                 .options.snow = opts) %dopar%
-    safe_fun(i)
-  
-  # stop cluster and progress bar
-  snow::stopCluster(cl)
-  
-  if (interactive() && show_progress) close(pb)
+  parallel_result <- furrr::future_map(in_paths, safe_fun, .progress = progress)
   
   
   res_transposed <- purrr::transpose(parallel_result)
@@ -110,7 +89,7 @@ jstor_convert_to_file <- function(in_paths, chunk_number, out_path, fun,
 #' many files easier:
 #' 
 #' - [purrr::safely()]
-#' - [foreach::foreach()]
+#' - [furrr::future_map()]
 #' - [readr::write_csv()]
 #' 
 #' When using one of the `find_*` functions, there should usually be no errors.
@@ -119,20 +98,20 @@ jstor_convert_to_file <- function(in_paths, chunk_number, out_path, fun,
 #' continue the process, and catch the error along the way.
 #' 
 #' If you have many files to import, you might benefit from executing the
-#' function in parallel. We use
-#' \code{\link[snow:snow-startstop]{snow::createCluster()}} to setup a cluster
-#' and
-#' then compute the results via `foreach` and \code{\%dopar\%}. The type of
-#' cluster is determined by `getClusterOption("type")`.
+#' function in parallel. We use futures for this to give you maximum 
+#' flexibility. By default the code is executed sequentially. If you want to
+#' run it in parallel, simply call [future::plan()] with
+#' [future::multiprocess()] as an argument before
+#' running `jst_import` or `jst_import_zip`. 
 #' 
 #' After importing all files, they are written to disk with
 #' [readr::write_csv()].
 #' 
 #' Since you might run out of memory when importing a large quantity of files,
-#' the files to import are split up into batches. Each batch is being treated
-#' separately, therefore for each batch multiple processes from
-#' \code{\link[snow:snow-startstop]{snow::createCluster()}} are
-#' spawned. For this reason, it is not recommended to have very small batches,
+#' you can split up the files to import  into batches. Each batch is being 
+#' treated separately, therefore for each batch multiple processes from
+#' [future::multiprocess()] are spawned, if you added this plan.
+#' For this reason, it is not recommended to have very small batches,
 #' as there is an overhead for starting and ending the processes. On the other
[…]
