+++ R/import_spec.R
@@ -152,7 +152,7 @@ check_env <- function(fun) {
 }
 
 
-walk_spec <- function(spec_df, chunk_number, n_batches, out_path, cores,
+walk_spec <- function(spec_df, chunk_number, n_batches, out_path,
                       show_progress, col_names) {
   message("Processing files for ", paste(unique(spec_df$meta_type), collapse = " and "),
           " with functions ", unique(spec_df$fun_names))
@@ -207,7 +207,7 @@ walk_spec <- function(spec_df, chunk_number, n_batches, out_path, cores,
       n_batches = n_batches, 
       out_path = out_paths_recycled, in_paths = in_paths_recycled,
       chunk_number = chunk_number_recycled,
-      fun = funs_recycled, cores = cores,
+      fun = funs_recycled,
       show_progress = show_progress, col_names = col_names
     ),
     jstor_convert_to_file
