### 2018-05-29 5c858dd
add future to suggests to avoid warning in tests
