### 2018-05-30 39470ae
refactor convert_page code for speed
