### 2018-05-30 ab32cdb
suppress warnings when reading xml from zip
