### 2018-05-31 4e9ddd1
expand augmentations #39
