+++ R/augmentations.R
@@ -26,25 +26,110 @@
 #'
 #' # get pages from page range
 #' jst_get_total_pages(page_range = "51 - 70, 350")
-jst_get_total_pages <- function(first_page = NULL, last_page = NULL,
-                             page_range = NULL) {
-  if (!is.null(page_range) && !is.na(page_range)) {
-    stopifnot(is.character(page_range))
+jst_get_total_pages <- function(first_page, last_page, page_range) {
+  ll <- list(first_page, last_page, page_range)
+  if (!identical(length(unique(lengths(ll))), 1L)) {
+    abort("All inputs must have the same length.")
+  }
+  
+  case_when(
+    is.na(first_page) & is.na(last_page) & is.na(page_range) ~ NA_real_,
+    !is.na(page_range) ~ wrap_parse_range(page_range),
+    is.na(page_range) & !is.na(first_page) & is.na(last_page) ~ NA_real_,
+    is.na(page_range) & 
+      !is.na(first_page) & 
+        !is.na(last_page) ~ as.numeric(last_page) - as.numeric(first_page) + 1
+  )
+}
+
+
+
+
+jst_add_total_pages <- function(meta_data, page_col = n_pages) {
+  page_col <- rlang::enquo(page_col)
+  dplyr::mutate(
+    meta_data,
+    !!page_col := jst_get_total_pages(first_page, last_page, page_range)
+  )
+}
+
+
+jst_unify_journal_id <- function(meta_data) {
+  meta_data %>%
+    dplyr::mutate(journal_id = case_when(is.na(journal_pub_id) ~ journal_jcode,
+                                         TRUE ~ journal_pub_id)) %>%
+    dplyr::select(-journal_pub_id, -journal_jcode, -journal_doi)
+}
+
+jst_clean_page <- function(page) {
+  # check if any has weird format like with AJS: AJSv104p126
+  is_complicated <- str_detect(page, "p")
+  
+  # do not change missing values
+  is_complicated[is.na(is_complicated)] <- FALSE
+  
+  page[is_complicated] <- stringr::str_extract(page[is_complicated], 
+                                               "(?<=p)\\d+")
+  
+  convert_page(page)
+}
+
+
+jst_augment <- function(meta_data) {
+  col_names <- names(meta_data)
+  
+  if (identical(col_names, names(article_cols$cols)) ||
+      identical(col_names, names(article_cols_old$cols))) {
+    # for journal articles
+    meta_data %>%
+      dplyr::mutate_at(vars("first_page", "last_page"), jst_clean_page) %>% 
+      jst_unify_journal_id() %>%
+      jst_add_total_pages()
     
-    first_page <- stringr::str_extract(page_range, "\\d+") %>% as.integer()
-    second_page <- stringr::str_extract(page_range, "(?<=-\\s?)\\d+") %>% 
-      as.integer()
-    
-    second_page - first_page + 1
-    
+  } else if (identical(col_names, names(book_cols$cols))) {
+    # for books
+    meta_data %>%
+      dplyr::mutate_at(vars("first_page"), jst_clean_page) 
   } else {
-    stopifnot(is.numeric(first_page) && is.numeric(last_page) || 
-                all(is.na(first_page)) || all(is.na(last_page)))
-    
-    dplyr::case_when(
-      !is.na(first_page) & is.na(last_page) ~ NA_real_,
-      !is.na(first_page) & !is.na(last_page) ~ last_page - first_page + 1,
-      TRUE ~ NA_real_
-    )
+    abort("Unknown meta_data type.")
   }
 }
+
[…]
