### 2018-05-31 6ad5b67
refactor parsing of page ranges by vectorizing it
