### 2018-05-31 ee3c094
add some quirks
