### 2018-05-31 85f51f5
document new functions
