+++ R/augmentations.R
@@ -1,3 +1,22 @@
+#' Clean data from DfR
+#' 
+#' This function takes data from [jst_get_article()] or [jst_get_book()] and
+#' applies helper functions for cleaning the data.
+#' 
+#' Data from DfR is inherently messy. For many examples see 
+#' `vignette("known-quirks", package = "jstor")`. `jst_augment()` is a
+#' convenience function that tries to deal with a few common tasks to
+#' clean the data.
+#' 
+#' For journal articles, it calles [jst_clean_page()] for first and last page,
+#' [jst_unify_journal_id()] and [jst_add_total_pages()].
+#' 
+#' @param meta_data Data which was processed via [jst_get_article()] or 
+#' [jst_get_book()]
+#' 
+#' @return A cleaned tibble.
+#' @seealso [jst_clean_page()] [jst_unify_journal_id()] [jst_add_total_pages()]
+#' 
 #' @export
 jst_augment <- function(meta_data) {
   col_names <- names(meta_data)
@@ -19,7 +38,23 @@ jst_augment <- function(meta_data) {
   }
 }
 
+#' Clean a character vector of pages
+#' 
+#' This function tries to convert character vectors into integers. This function
+#' should not be called on page ranges.
+#' 
+#' @param page A character vector for pages.
 #' @export
+#' @examples 
+#' jst_clean_page("2")
+#' 
+#' # anything that is not a digit gets removed
+#' jst_clean_page("A2-")
+#' 
+#' # a weird format from the American Journal of Sociology is convered correctly
+#' jst_clean_page("AJSv104p126")
+#' # this is done by searching for "p", and if it is found, extracting the
+#' content after "p".
 jst_clean_page <- function(page) {
   # check if any has weird format like with AJS: AJSv104p126
   is_complicated <- str_detect(page, "p")
@@ -33,6 +68,13 @@ jst_clean_page <- function(page) {
   convert_page(page)
 }
 
+#' Add total count of pages
+#' 
+#' This function adds a column with the total count of pages. It calls 
+#' [jst_get_total_pages()] which does the work.
+#' 
+#' @param meta_data Data which was processed via [jst_get_article()].
+#' @param page_col The name of the new column with total count of pages.
 #' @export
 jst_add_total_pages <- function(meta_data, page_col = n_pages) {
   page_col <- rlang::enquo(page_col)
@@ -48,12 +90,22 @@ jst_add_total_pages <- function(meta_data, page_col = n_pages) {
 #' This function is a simple helper to calculate the total number of pages of
 #' an article.
 #' 
-#' This function deals with three cases:
+#' This function deals with four cases:
 #' 
+#' - if all three arguments are missing, NA is returned.
 #' - if page_range is supplied, the number of pages is calculated from it.
-#' - if only the first page is supplied, NA is returned
+#' - if only the first page is supplied, NA is returned.
 #' - if first and last page are supplied, the number of pages is calculated as
-#' `last_page - first_page + 1`
+#' `last_page - first_page + 1`.
+#' 
+#' The algorithm to parse page ranges works as follows: A typical page range is
+#' `1-10, 200` where the article starts at page 1, ends at page 10, and has an
+#' erratum at page 200. For this case, the range is calculated as
+#' `range + single_page`, as in`(10 - 1 + 1) + 1 = 11`. Sometimes multiple 
+#' ranges are given: `1-10, 11-20`. For those cases all ranges are summed:
[…]
