+++ man/jst_add_total_pages.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/augmentations.R
+\name{jst_add_total_pages}
+\alias{jst_add_total_pages}
+\title{Add total count of pages}
+\usage{
+jst_add_total_pages(meta_data, page_col = n_pages)
+}
+\arguments{
+\item{meta_data}{Data which was processed via \code{\link[=jst_get_article]{jst_get_article()}}.}
+
+\item{page_col}{The name of the new column with total count of pages.}
+}
+\description{
+This function adds a column with the total count of pages. It calls
+\code{\link[=jst_get_total_pages]{jst_get_total_pages()}} which does the work.
+}
