+++ man/jst_clean_page.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/augmentations.R
+\name{jst_clean_page}
+\alias{jst_clean_page}
+\title{Clean a character vector of pages}
+\usage{
+jst_clean_page(page)
+}
+\arguments{
+\item{page}{A character vector for pages.}
+}
+\description{
+This function tries to convert character vectors into integers. This function
+should not be called on page ranges.
+}
+\examples{
+jst_clean_page("2")
+
+# anything that is not a digit gets removed
+jst_clean_page("A2-")
+
+# a weird format from the American Journal of Sociology is convered correctly
+jst_clean_page("AJSv104p126")
+# this is done by searching for "p", and if it is found, extracting the
+content after "p".
+}

[…]
