### 2018-05-31 802fd66
don't load packages when testing
