### 2018-05-31 fbfd3c5
fix check issues
