### 2018-05-31 08af2de
reformat global variables
