### 2018-05-31 33c0bf8
do not augment books since there is nothing to augment
