### 2018-05-31 67e1514
add cross-links in documentation
