+++ man/jst_augment.Rd
@@ -7,14 +7,13 @@
 jst_augment(meta_data)
 }
 \arguments{
-\item{meta_data}{Data which was processed via \code{\link[=jst_get_article]{jst_get_article()}} or
-\code{\link[=jst_get_book]{jst_get_book()}}}
+\item{meta_data}{Data which was processed via \code{\link[=jst_get_article]{jst_get_article()}}.}
 }
 \value{
 A cleaned tibble.
 }
 \description{
-This function takes data from \code{\link[=jst_get_article]{jst_get_article()}} or \code{\link[=jst_get_book]{jst_get_book()}} and
+This function takes data from \code{\link[=jst_get_article]{jst_get_article()}} and
 applies helper functions for cleaning the data.
 }
 \details{
@@ -23,9 +22,10 @@ Data from DfR is inherently messy. For many examples see
 convenience function that tries to deal with a few common tasks to
 clean the data.
 
-For journal articles, it calles \code{\link[=jst_clean_page]{jst_clean_page()}} for first and last page,
-\code{\link[=jst_unify_journal_id]{jst_unify_journal_id()}} and \code{\link[=jst_add_total_pages]{jst_add_total_pages()}}.
+For journal articles, it calles \code{\link[=jst_clean_page]{jst_clean_page()}} to convert first and last
+page, \code{\link[=jst_unify_journal_id]{jst_unify_journal_id()}} and \code{\link[=jst_add_total_pages]{jst_add_total_pages()}}.
 }
 \seealso{
 \code{\link[=jst_clean_page]{jst_clean_page()}} \code{\link[=jst_unify_journal_id]{jst_unify_journal_id()}} \code{\link[=jst_add_total_pages]{jst_add_total_pages()}}
+\code{\link[=jst_get_total_pages]{jst_get_total_pages()}}
 }
