### 2018-05-31 03eb355
pub_year should be re-imported as integer for articles
