### 2018-05-31 f0c9d95
remove global variable since namespace was fixed
