### 2018-05-31 556015e
use augmentations in case study
