### 2018-05-31 2688381
remove exploration of missing journal ids and use jst_augment instead
