### 2018-05-31 cc72f51
use menu in pkgdown again
