### 2018-05-31 271da3f
improve doc wording
