+++ DESCRIPTION
@@ -29,6 +29,7 @@ Suggests:
     knitr,
     rmarkdown,
     future
+Remotes: r-lib/rlang
 BugReports: https://github.com/tklebel/jstor/issues
 URL: https://github.com/tklebel/jstor, https://tklebel.github.io/jstor/
 RoxygenNote: 6.0.1
