### 2018-05-31 67fae01
use abort instead of stop in some places
