### 2018-06-01 f638906
expand vignette on known quirks
closes #44
