+++ vignettes/known-quirks.Rmd
@@ -156,7 +156,9 @@ Identifiers for the journal usually appear in three columns:
 - `journal_pub_id`
 
 ```{r, results='asis'}
-jst_get_article(jst_example("sample_with_references.xml")) %>% knitr::kable()
+sample_article <- jst_get_article(jst_example("sample_with_references.xml")) 
+
+knitr::kable(sample_article)
 ```
 
 From my samples, it seems that the information in `journal_pub_id` is 
@@ -171,7 +173,7 @@ general information about the respective journals, which are available from
 `jst_get_journal_overview()`:
 
 ```{r}
-jst_get_article(jst_example("sample_with_references.xml")) %>% 
+sample_article %>% 
   jst_unify_journal_id() %>% 
   left_join(jst_get_journal_overview()) %>% 
   tidyr::gather(variable, value) %>% 
@@ -194,3 +196,17 @@ since they are calculated on all reviews in the issue, irrespective of
 whether actually all reviews of the given issue are in the sample or not.
 Alternatively, one could group by issues, and only take one set of ngrams
 per issue.
+
+## Language codes
+Information on langues is not consistent. For the sample article, `language` is
+`eng`. 
+
+```{r}
+sample_article %>% 
+  pull(language)
+```
+
+In other cases it might be `en`. It is thus advisable to take a quick look at
+different variants via `distinct(meta_data, language)` or 
+`count(meta_data, language)`. 
+
