### 2018-06-01 dcf4d04
add jst_preview_zip to the case study
