+++ NEWS.md
@@ -1,48 +1,6 @@
 # jstor 0.3.0
 
-# jstor 0.2.6.9000
-
-## Major changes
-There is a new set of functions which lets you directly import files from 
-.zip-archives: `jst_import_zip()` and `jst_define_import()`.
-
-In the following example, we have a zip-archive from DfR and want to import
-metadata on books and articles. For all articles we want to apply
-`find_article()` and `find_authors()`, for books only `find_book()`, and we want
-to read unigrams (ngram1).
-
-First we specify what we want, and then we apply it to our zip-archive.
-```r
-# specify definition
-import_spec <- jst_define_import(article = c(find_article, find_authors),
-                                 book = find_book,
-                                 ngram1 = jst_read_ngram)
-
-# apply definition to archive
-jst_import_zip("zip_archive.zip",
-               import_spec = import_spec,
-               out_file = "out_path")
-```
-
-If the archive contains also research reports, pamphlets or other ngrams, they
-will not be imported. We could however change our specification, if we wanted
-to import all kinds of ngrams (given that we originally requested them from
-DfR):
-
-```r
-# import multiple forms of ngrams
-import_spec <- jst_define_import(article = c(find_article, find_authors),
-                                 book = find_book,
-                                 ngram1 = jst_read_ngram,
-                                 ngram2 = jst_read_ngram,
-                                 ngram3 = jst_read_ngram)
-```
-
-Before importing all files from a zip-archive, you can get a quick overview with
-`jst_preview_zip()`.
-
-
-## Other changes
+## Breaking changes
 `jst_import` and `jst_import_zip` now use futures as a backend for parallel 
 processing. This makes internals more compact and reduces dependencies. 
 Furthermore this reduces the number of arguments, since the argument `cores` 
@@ -58,37 +16,94 @@ jst_import_zip("zip-archive.zip",
 ```
 If you want to terminate the proceses, at least on *nix-systems you need to kill
 them manually (once again).
-* The new `vignette("known-quirks")` lists common problems with data from
+
+* All functions have been renamed to use a unified naming scheme: `jst_*`.
+The former group of `find_*` functions is now called `jst_get_*`, as in
+`jst_get_article()`. The previous functions have been deprecated and will be 
+removed before submission to CRAN.
+* The unique identifier for matching across files has been renamed to 
+`file_name`, and the corresponding helper to get this file name from
+`get_basename` to `jst_get_file_name`.
+
+## Importing data directly from zip-files
+There is a new set of functions which lets you directly import files from 
+.zip-archives: `jst_import_zip()` and `jst_define_import()`.
+
+In the following example, we have a zip-archive from DfR and want to import
+metadata on books and articles. For all articles we want to apply
+`jst_get_article()` and `jst_get_authors()`, for books only `jst_get_book()`, 
+and we want to read unigrams (ngram1).
+
+First we specify what we want, and then we apply it to our zip-archive:
+```r
+# specify definition
+import_spec <- jst_define_import(article = c(jst_get_article, jst_get_authors),
+                                 book = jst_get_book,
+                                 ngram1 = jst_get_ngram)
+
+# apply definition to archive
+jst_import_zip("zip_archive.zip",
+               import_spec = import_spec,
+               out_file = "out_path")
+```
+
+If the archive contains also research reports, pamphlets or other ngrams, they
+will not be imported. We could however change our specification, if we wanted
+to import all kinds of ngrams (given that we originally requested them from
+DfR):
+
+```r
+# import multiple forms of ngrams
+import_spec <- jst_define_import(article = c(jst_get_article, jst_get_authors),
+                                 book = jst_get_book,
+                                 ngram1 = jst_get_ngram,
[…]
