### 2018-06-05 ff0cdf0
fix error in README
