### 2018-06-05 086c61d
fix doc error for re-import
