+++ R/re-import.R
@@ -38,10 +38,10 @@
 #' tmp <- tempdir()
 #' 
 #' # find multiple files
-#' file_list <- rep(jstor_example("sample_with_references.xml"), 2)
+#' file_list <- rep(jst_example("sample_with_references.xml"), 2)
 #'
 #' # convert and write to file
-#' jst_import(file_list, "article", out_path = tmp, .f = find_article,
+#' jst_import(file_list, "article", out_path = tmp, .f = jst_get_article,
 #'              n_batches = 2, show_progress = FALSE)
 #'              
 #' # combine outputs
