+++ NEWS.md
@@ -1,3 +1,5 @@
+# jstor 0.3.0.9000
+
 # jstor 0.3.0
 
 ## Breaking changes
