### 2018-06-05 #189 @elinw
I'm sorry I'm slow on replying but this seems good. I love the import from zip.  I have some things to focus on right now but by the end of the week I'm planning to have updated to the current development branch and try things out.   I will also get you an example article.
