### 2018-06-06 #189 @elinw
25074331
 basename_id journal_doi journal_jcode journal_pub_id article_doi article_pub_id article_jcode     article_type
1    25074331          NA     jbusiethi                                                25074331 research-article
                                                                  article_title volume issue language pub_day
1 Ethics Education in the Workplace: An Effective Tool to Combat Employee Theft     26     2      eng       1
  pub_month pub_year first_page last_page
1         7     2000         89       100
Has
21646 Greengard, Samuel: April 1993, 'Theft Control\nStarts with HR Strategies', Personnel Journal, p. 85.
21647 Greengard, Samuel: April 1993, 'Theft Control\nStarts with HR Strategies', Personnel Journal, p. 85.
21648 Greengard, Samuel: April 1993, 'Theft Control\nStarts with HR Strategies', Personnel Journal, p. 85.
21649 Greengard, Samuel: April 1993, 'Theft Control\nStarts with HR Strategies', Personnel Journal, p. 85.
21650 Greengard, Samuel: April 1993, 'Theft Control\nStarts with HR Strategies', Personnel Journal, p. 85.

I actually pulled the article and it's true, they cited the same page multiple times.  It was striking because it made Greengard the most cited author but I had never heard of him.
