### 2018-06-07 9a2147c
fix news typos
