### 2018-06-07 cf1b405
expand known quirks a bit
