### 2018-06-07 #189 @tklebel
Thanks @elinw for the example. I added a few sentences about issues with references [in the vignette](https://tklebel.github.io/jstor/articles/known-quirks.html#incorrectodd-references) and linked the article as an example.
