### 2018-06-07 #189 @elinw
I actually don't think that the article is "wrong" it's just following the citation style of the journal it is published in.  Other journals probably would have used Ibid and op. cit. or similar.
