### 2018-06-10 8a3d487
remove certatinty of statement about citation artifact
