### 2018-06-27 #189 @karthik
Hi @tklebel Apologies for the delay. I'm just waiting on @elinw to sign off and with that we can finish up.
Elin: Do you have any further unresolved issue with the master release of this effort? If not could you indicate sign off so we can proceed with acceptance? 🙏
