### 2018-06-27 #189 @elinw
Yes, I am good with these changes, the last of which are mainly just polishing documentation etc.  This is going to be really useful and encourage the use of JfR.   I noticed the change from stop() to abort() and read about the differences, my suggestion is that you make that rlang::abort() explicitly, which you'll need to do for CRAN anyway.

Great job!
