### 2018-06-27 #189 @tklebel
@elinw I'm actually not sure if I really prefer `abort` over `stop` in general. I used it to remove the calls to `.call = FALSE`. But since I import it via `importFrom(rlang,abort)` in NAMESPACE, this should be fine for CRAN, right? If not, I will of course change it before submitting.

Apart from that, I am very happy for the thumbs up! This comes just in time before useR! :sei
