### 2018-06-27 59a42e2
add rOpenSci footer and update links
