+++ README.Rmd
@@ -14,9 +14,9 @@ knitr::opts_chunk$set(
 ```
 # jstor
 
-[![Travis build status](https://travis-ci.org/tklebel/jstor.svg?branch=master)](https://travis-ci.org/tklebel/jstor)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/tklebel/jstor?branch=master&svg=true)](https://ci.appveyor.com/project/tklebel/jstor)
-[![Coverage status](https://codecov.io/gh/tklebel/jstor/branch/master/graph/badge.svg)](https://codecov.io/github/tklebel/jstor?branch=master)
+[![Travis build status](https://travis-ci.org/ropensci/jstor.svg?branch=master)](https://travis-ci.org/ropensci/jstor)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/jstor?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/jstor)
+[![Coverage status](https://codecov.io/gh/ropensci/jstor/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/jstor?branch=master)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![CRAN status](http://www.r-pkg.org/badges/version/jstor)](https://cran.r-project.org/package=jstor)
 <!---[![CRAN\_Download\_Badge](http://cranlogs.r-pkg.org/badges/jstor)](https://CRAN.R-project.org/package=jstor)--->
@@ -55,7 +55,7 @@ You can install `jstor` from GitHub with:
 
 ```{r gh-installation, eval = FALSE}
 # install.packages("devtools")
-devtools::install_github("tklebel/jstor")
+devtools::install_github("ropensci/jstor")
 ```
 
 ## Usage
@@ -170,3 +170,6 @@ project number "P 29211 Einzelprojekte".
 
 Some internal functions regarding file paths and example files were adapted from
 the package `readr`.
+
+[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
+
