### 2018-06-28 c95a019
fix appveyor link
