+++ README.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(
 # jstor
 
 [![Travis build status](https://travis-ci.org/ropensci/jstor.svg?branch=master)](https://travis-ci.org/ropensci/jstor)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/jstor?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/jstor)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/jstor?branch=master&svg=true)](https://ci.appveyor.com/project/tklebel/jstor)
 [![Coverage status](https://codecov.io/gh/ropensci/jstor/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/jstor?branch=master)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![CRAN status](http://www.r-pkg.org/badges/version/jstor)](https://cran.r-project.org/package=jstor)
