+++ README.Rmd
@@ -12,7 +12,10 @@ knitr::opts_chunk$set(
   out.width = "100%"
 )
 ```
-# jstor
+# jstor: Import and Analyse Data from Scientific Articles
+
+**Author:** [Thomas Klebel](https://thomasklebel.eu) <br>
+**License:** [GPL v3.0](https://www.gnu.org/licenses/gpl-3.0.en.html)
 
 [![Travis build status](https://travis-ci.org/ropensci/jstor.svg?branch=master)](https://travis-ci.org/ropensci/jstor)
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/jstor?branch=master&svg=true)](https://ci.appveyor.com/project/tklebel/jstor)
