### 2018-07-07 2a96cd5
remove argument `new_col` which needs devel rlang
