### 2018-07-08 e5ba6d3
fix link in news
