### 2018-07-08 83d20f0
update vignette for batch import #56
