+++ vignettes/automating_file_import.Rmd
@@ -16,26 +16,141 @@ knitr::opts_chunk$set(
 )
 ```
 
+```{r}
+library(jstor)
+library(purrr)
+library(stringr)
+```
+
+
 # Intro
 The `find_*` functions from `jstor` all work on a single file. Data from DfR
 however contains many single files, from up to 25,000 when using the self-service
 functions, up to several hundreds of thousands of files when requesting a large
 dataset via an agreement.
 
-In this vignette I will introduce _one_ way to deal with this large amount of
-files. I start with how to list all corresponding files and follow up with some
-utility functions implemented in `jstor`. Possibly there are other ways one could
-import all those files, but this is how I did it four our research.
+Currently `jstor` offers two implementations to import many files: 
+`jst_import_zip()` and `jst_import()`. The first one lets you import data directly
+from the zip archive, the second one works for file paths, so you need to unzip
+the archive first.
+
+I will first introduce `jst_import_zip()` and discuss the approach of with
+`jst_import()` afterwards.
+
+# Importing data directly from zip-archives
+Unpacking and working with many files directly is unpractical for at least three
+reasons: 
+
+1. If you unzip the archive, the single files will occupy a lot more disk space
+than the single archive.
+2. Before you can import the files, you need to list them  via `list.files` or
+`system("find...")` on UNIX. Depending on the size of your data, this can take
+some time.
+3. There might be different types of data in your sample (journal articles, book
+chapters, etc.). You need to manage the paths in order to apply the correct
+functions to them.
+
+Importing directly from the zip archive simplifies all those tasks with a
+single function: `jst_import_zip()`. For the following demonstration, we will
+use a small sample archive that comes with the package.
+
+As a first step, we should take a look at the archive and its content. This
+is made easy with `jst_preview_zip():
+
+```{r}
+jst_preview_zip(jst_example("pseudo_dfr.zip"))
+```
+
+We can see that we have a simple archive with three metadata files and one 
+ngram file. Before we can use `jst_import_zip()`, we first need to think about,
+what we acutally want to import: all of the data, or just parts? What kind of
+data do we want to extract from articles, books and pamphlets? We can specify
+this via `jst_define_import()`:
+
+```{r}
+import_spec <- jst_define_import(article = c(jst_get_article, jst_get_authors),
+                                 book = jst_get_book,
+                                 ngram1 = jst_get_ngram)
+```
+
+In this case, we want to import data on articles (the standard metadata plus
+information on the authors), general data on books and unigrams (ngram1). This
+specification can then be used with `jst_import_zip()`:
+
+```{r}
+# set up a temporary folder for output files
+tmp <- tempdir()
+
+# extract the content and write output to disk
+jst_import_zip(jst_example("pseudo_dfr.zip"),
+               import_spec = import_spec,
+               out_file = "my_test",
+               out_path = tmp)
+
+```
+
+We can take a look at the files within `tmp` with `list.files()`:
+```{r}
+list.files(tmp, pattern = "csv")
+```
+
+As you can see, `jst_import_zip()` automatically creates file names based on
+the string you supplied to `out_file` to delianate the different types of output.
+
+If we want to re-import the data for further analysis, we can either use 
+functions like `readr::read_csv()`, or a small helper function from the package
+which determines and sets the column types correctly:
+
+```{r}
+jst_re_import(paste0(tmp, "/my_test_journal_article_jst_get_article-1.csv")) %>% 
+  knitr::kable()
+```
+
+A side note on ngrams: For larger archives, importing all ngrams takes a very
+long time. It is thus advisable to only import ngrams for articles which you
[…]
