### 2018-07-08 982aedc
don't run multiple cores in vignette
