### 2018-07-08 8a96e17
proofread vignette
closes #56
