### 2018-07-08 ca63b65
add cran-comments
