+++ cran-comments.md
@@ -1,0 +1,10 @@
+## Test environments
+* local OS X install, R 3.5.0
+* ubuntu 14.04 (on travis-ci), devel and release
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* This is a new release.
