+++ appveyor.yml
@@ -15,8 +15,8 @@ cache:
 
 # Adapt as necessary starting from here
 
-environment:
-  USE_RTOOLS: true
+#environment:
+#  USE_RTOOLS: true
 
 build_script:
   - travis-tool.sh install_deps
