### 2018-07-08 3816e79
finally fix links for ci properly
