### 2018-07-08 b448cc2
use different email for cran submission
