+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: jstor
 Title: Read Data from JSTOR/DfR
 Version: 0.3.0.9000
-Authors@R: person("Thomas", "Klebel", email = "thomas.klebel@uni-graz.at", 
+Authors@R: person("Thomas", "Klebel", email = "t.klebel@hotmail.com", 
                   role = c("aut", "cre"))
 Description: Functions and helpers to import metadata and full-texts delivered
     by Data for Research (DfR) by JSTOR. 
