### 2018-07-08 2340226
fix old functions in intro vignette
