### 2018-07-08 5d23992
and revert again to old link for appveyor
