+++ DESCRIPTION
@@ -3,8 +3,8 @@ Title: Read Data from JSTOR/DfR
 Version: 0.3.1
 Authors@R: person("Thomas", "Klebel", email = "t.klebel@hotmail.com", 
                   role = c("aut", "cre"))
-Description: Functions and helpers to import metadata and full-texts delivered
-    by Data for Research (DfR) by JSTOR. 
+Description: Functions and helpers to import metadata, ngrams and full-texts 
+    delivered by Data for Research by JSTOR. 
 Depends: R (>= 3.1)
 License: GPL-3 | file LICENSE
 Encoding: UTF-8
