### 2018-07-14 50a2351
include install info for CRAN
