### 2018-07-15 8a00369
replace paste0 with file.path to ensure we have valid paths
this is in response to CRAN archiving the package, because "test-writing-to-file" was incorrect in line 63
