+++ tests/testthat/test-zip.R
@@ -45,8 +45,8 @@ test_that("importing from zip works", {
   
   # test for the content of the files
   test_expected_zip <- function(expected_file) {
-    expect_identical(read_csv(paste0("testfiles/", expected_file)),
-                     read_csv(paste0(test_dir, "/", expected_file)))
+    expect_identical(read_csv(file.path("testfiles", expected_file)),
+                     read_csv(file.path(test_dir, expected_file)))
   }
   expected_files %>% 
     purrr::map(test_expected_zip)
