+++ vignettes/automating_file_import.Rmd
@@ -103,7 +103,7 @@ which determines and sets the column types correctly:
 
 ```{r}
 jst_re_import(
-  paste0(tmp, "/my_test_journal_article_jst_get_article-1.csv")
+  file.path(tmp, "my_test_journal_article_jst_get_article-1.csv")
 ) %>% 
   knitr::kable()
 ```
