+++ NEWS.md
@@ -1,3 +1,5 @@
+# jstor 0.3.2
+
 # jstor 0.3.1.9000
 
 # jstor 0.3.1
