+++ NEWS.md
@@ -1,6 +1,7 @@
 # jstor 0.3.2
+This is a hotfix to resolve an issue with writing to other directories than
+temporary folders during tests, which should not have happend.
 
-# jstor 0.3.1.9000
 
 # jstor 0.3.1
 
