### 2018-07-15 25844ec
change news slightly
