### 2018-07-15 f390332
update cran-comments
