### 2018-08-07 e2a5c75
change cran badge from monthly to total
closes #67
