+++ DESCRIPTION
@@ -31,6 +31,6 @@ Suggests:
     future
 BugReports: https://github.com/ropensci/jstor/issues
 URL: https://github.com/ropensci/jstor, https://ropensci.github.io/jstor/
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
 Roxygen: list(markdown = TRUE)
 VignetteBuilder: knitr
