### 2018-08-07 fe3ed11
fix typo and add examples
