### 2018-08-07 dc1cd6b
fix prefix of functions in various places
(jstor_ -> jst_)
