+++ R/re-import.R
@@ -1,10 +1,10 @@
 #' Combine outputs from converted files
 #' 
 #' `jst_combine_outputs()` helps you to manage the multitude of files you might
-#' receive after running [jstor_import()] or [jst_import_zip()] with more than
+#' receive after running [jst_import()] or [jst_import_zip()] with more than
 #' one batch. 
 #' 
-#' Splitting the output of [jstor_import()] or [jst_import_zip()] might be done
+#' Splitting the output of [jst_import()] or [jst_import_zip()] might be done
 #' for multiple reasons, but in the end you possibly want to combine all outputs
 #' into one file/data.frame. This function makes a few assumptions in order to
 #' combine files: 
@@ -15,12 +15,12 @@
 #'   If you want to combine `foo-1.csv` and `foo-2.csv`, the combined file will
 #'   be `combined_foo.csv`.
 #' - The directory only contains files which were imported via 
-#'   [jstor_import()] or [jst_import_zip()]. If the directory contains other
+#'   [jst_import()] or [jst_import_zip()]. If the directory contains other
 #'   `.csv` files, you should supply a character vector with paths to only those
 #'   files, which you want to import.
 #' 
 #' @param path A path to a directory, containing .csv-files from
-#'  [jstor_import()] or [jst_import_zip()], or a vector of files which are to be
+#'  [jst_import()] or [jst_import_zip()], or a vector of files which are to be
 #'  imported.
 #' @param write_to_file Should combined data be written to a file?
 #' @param out_path A directory where to write the combined files. If no
@@ -137,7 +137,7 @@ jst_combine_outputs <- function(path, write_to_file = TRUE,
 #' Re-import files 
 #' 
 #' `jst_re_import()` lets you re-import a file which was exported via 
-#' [jstor_import()] or [jst_import_zip()].
+#' [jst_import()] or [jst_import_zip()].
 #' 
 #' When attempting to re-import, a heuristic is applied. If the file has column
 #' names which match the names from any of the `find_*` functions, the file
@@ -158,7 +158,7 @@ jst_re_import <- function(file, warn = TRUE) {
   file <- check_path(file)
   
   if (!identical(tools::file_ext(file), "csv")) {
-    abort(paste("Only .csv-files which were generated by `jstor_import` or",
+    abort(paste("Only .csv-files which were generated by `jst_import` or",
                 "`jst_import_zip` can be re-imported."))
   }
   
