+++ man/jst_import.Rd
@@ -107,7 +107,7 @@ jst_import(meta_files, out_file = "imported_metadata", .f = jst_get_article,
 imports <- jst_define_import(article = c(jst_get_article, jst_get_authors))
 
 # convert the files to .csv
-jstor_import_zip("my_archive.zip", out_file = "my_out_file", 
+jst_import_zip("my_archive.zip", out_file = "my_out_file", 
                  import_spec = imports)
 } 
 }
