### 2018-08-07 f4bc8aa
rework paper for JOSS
