+++ paper.bib
@@ -25,12 +25,12 @@
 	file = {Full Text PDF:/Users/tklebel/Documents/zotero/storage/BRIWHQBA/Roberts - 2016 - Introduction to the Virtual Issue Recent Innovati.pdf:application/pdf;Snapshot:/Users/tklebel/Documents/zotero/storage/F4PLIM9P/8CFCA019C25D82254EE0CAC28F3F650E.html:text/html}
 }
 
- @Manual{r_core,
+  @Manual{r_core,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
-    year = {2017},
+    year = {2018},
     url = {https://www.R-project.org/},
   }
   
@@ -39,6 +39,6 @@
     title = {jstor: Read Data from JSTOR/DfR},
     author = {Thomas Klebel},
     year = {2018},
-    note = {R package version 0.2.5.9000},
-    url = {https://github.com/tklebel/jstor},
+    note = {https://github.com/ropensci/jstor, https://ropensci.github.io/jstor/},
+    url = {https://ropensci.github.io/jstor/}
   }
