### 2018-08-07 178b41e
update appveyor badge
