+++ README.Rmd
@@ -18,7 +18,7 @@ knitr::opts_chunk$set(
 **License:** [GPL v3.0](https://www.gnu.org/licenses/gpl-3.0.en.html)
 
 [![Travis build status](https://travis-ci.org/ropensci/jstor.svg?branch=master)](https://travis-ci.org/ropensci/jstor)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/qtxvd3tugf2l589c?svg=true)](https://ci.appveyor.com/project/tklebel/jstor)
+[![AppVeyorBuild status](https://ci.appveyor.com/api/projects/status/sry2gtwam7qyfw6l?svg=true)](https://ci.appveyor.com/project/tklebel/jstor)
 [![Coverage status](https://codecov.io/gh/ropensci/jstor/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/jstor?branch=master)
 [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![CRAN status](http://www.r-pkg.org/badges/version/jstor)](https://cran.r-project.org/package=jstor)
