### 2018-08-07 fada33f
refactor paper
