### 2018-08-07 4481e2e
further tweak paper
