+++ paper.md
@@ -12,6 +12,9 @@ bibliography: paper.bib
 tags:
  - JSTOR
  - DfR
+ - Data for Research
+ - scientometrics
+ - bibliometrics
  - text mining
  - text analysis
  - citation analysis
@@ -23,23 +26,22 @@ affiliations:
 # Summary
 The interest in text as data has has seen a sharp increase in the 
 past few years, mostly due to the advent of methods for automated text analysis.
-Every analysis of text needs some 
-data and hence data availability is an important issue.
-For researchers
-interested in the study of science and the scholarly literature in particular,
-the archival content of [JSTOR](http://www.jstor.org) offers a rich and diverse
-set of primary sources like research articles or book chapters. The service 
+At the same time, researches within the field of scientometrics have analysed
+citations and other aspects of the scholarly literature with great sophistication.
+The archival content of [JSTOR](http://www.jstor.org) offers a rich and diverse
+set of primary sources like research articles or book chapters for both 
+approaches. 
 [Data for Research (DfR)](http://www.jstor.org/dfr/) by JSTOR gives all 
 researchers, regardless of whether they have access to JSTOR or not, the
-opportunity to analyse this rich dataset, by delivering metadata,
+opportunity to analyse metadata,
 n-grams and, upon special request, full-text materials about all available
 articles and books from JSTOR. The package `jstor` [@jstor] helps in
 analysing these datasets by enabling researchers to easily import the metadata
-to R [@r_core], for which no other integrated solution exists to date.
+to R [@r_core], a task, for which no other integrated solution exists to date.
 
 The metadata from DfR
 can either be analysed on their own or be used in conjunction with n-grams
-or full-text-data. Commonly, metadata from DfR include information
+or full-text data. Commonly, metadata from DfR include information
 on the articles' authors, the title, journal, date of publishing, and quite
 frequently all footnotes and references. All this information can be of interest
 for specific research questions. For the analysis of n-grams or full-texts,
@@ -48,7 +50,7 @@ allow the researchers to
 filter articles based on specific journals, the dates of publication, the
 authors, keywords in titles and other aspects.
 
-`jstor` provides functions for three main tasks:
+`jstor` provides functions for three main tasks within the research process:
 
 - Importing different parts of metadata, either from XML-files or directly from
 the .zip-archive provided by DfR.
@@ -59,7 +61,7 @@ cleaning page numbers.
 
 Full documentation of `jstor`, including a comprehensive 
 case study about analysing 
-n-grams, is available at 
+n-grams from DfR, is available at 
 https://ropensci.github.io/jstor/. The package can be obtained from 
 CRAN (https://CRAN.R-project.org/package=jstor)
 or from GitHub (https://github.com/ropensci/jstor). 
