### 2018-08-08 565ae2e
mention in case study that package is on CRAN
